### 2025-03-21 #7938 @editorialbot
[PRE REVIEW]: pangeo-fish: A Python package for studying fish movement using biologging and earth science data
**Submitting author:** @keewis (<a href="http://orcid.org/0000-0002-4254-8002">Justus Magin</a>)
**Repository:** https://github.com/pangeo-fish/pangeo-fish
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 2025.03.0
