### 2025-03-24 0c4e606
Link to report issues (#178)
* updated CI's requirements.

* added a link to report issues to Github in docs.
