### 2025-03-24 7b61605
fixed args in nbstripout + added back language info in notebook pangeo-fish.
