+++ .pre-commit-config.yaml
@@ -35,8 +35,7 @@ repos:
     hooks:
       - id: nbstripout
         args:
-          - "--extra-keys=metadata.kernelspec"
-          - "metadata.language_info.version"
+          - "--extra-keys=metadata.kernelspec metadata.language_info.version"
 
   - repo: https://github.com/rbubley/mirrors-prettier
     rev: v3.5.3
