### 2025-03-24 b532799
added upper bound version to xarray.
