### 2025-03-24 4215503
Healpixify renaming (#179)
* updated CI's requirements.

* renamed xarray-healpy imports/use to xhealpixify.
