+++ docs/environment.yaml
@@ -38,5 +38,6 @@ dependencies:
   - pip:
       - nbsphinx_link>=1.3.1
       - xdggs
-      - xarray-healpy @ git+https://github.com/iaocea/xarray-healpy.git@0ffca6058f4008f4f22f076e2d60787fcf32ac82
+      # - xarray-healpy @ git+https://github.com/iaocea/xarray-healpy.git@0ffca6058f4008f4f22f076e2d60787fcf32ac82
+      - xhealpixify
       - -e ../.
