+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
       - id: blackdoc-autoupdate-black
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.0
+    rev: v0.11.2
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
@@ -58,6 +58,6 @@ repos:
         args: ["--no-schema"]
 
   - repo: https://github.com/abravalheri/validate-pyproject
-    rev: v0.24
+    rev: v0.24.1
     hooks:
       - id: validate-pyproject
