### 2025-03-25 #7938 @QuentinMaz
I will deal with the [DOI-related issues](https://github.com/openjournals/joss-reviews/issues/7938#issuecomment-2743803212).
