+++ publication/paper.bib
@@ -1,12 +1,17 @@
 @article{de2019new,
-  title={New insights into behavioural ecology of European seabass off the West Coast of France: implications at local and population scales},
-  author={{de Pontual}, Helene and Lalire, Maxime and Fablet, Ronan and Laspougeas, Claire and Garren, Fran{\c{c}}ois and Martin, St{\'e}phane and Drogou, Mickael and Woillez, Mathieu},
-  journal={ICES Journal of Marine Science},
-  volume={76},
-  number={2},
-  pages={501--515},
-  year={2019},
-  publisher={Oxford University Press}
+    author = {de Pontual, Hélène and Lalire, Maxime and Fablet, Ronan and Laspougeas, Claire and Garren, François and Martin, Stéphane and Drogou, Mickael and Woillez, Mathieu},
+    title = {New insights into behavioural ecology of European seabass off the West Coast of France: implications at local and population scales},
+    journal = {ICES Journal of Marine Science},
+    volume = {76},
+    number = {2},
+    pages = {501-515},
+    year = {2018},
+    month = {07},
+    abstract = {From 2010 to 2012, 246 data storage tags were deployed on European seabass in the Iroise Natural Marine Park, a marine protected area (MPA) off west Brittany, France. A return rate of 14.6\% associated with long time series of data provided new information on fish ecology (e.g. maximum experienced depth greater than 225 m, temperature range 6.80–21.87°C). Depth and temperature series were used to infer individual migration using an innovative hidden Markov model (HMM) especially developed for seabass geolocation. Reconstructed fish tracks revealed that seabass is a partially migratory species, as individuals exhibited either long-distance migrations towards the Bay of Biscay or the Celtic Sea, or residency behaviour in the Iroise Sea. Fidelity to summer feeding areas and to winter spawing areas was demonstrated. These results suggest that the population is spatially structured. The Iroise Sea is likely a mixing zone for different stocks or sub-populations, and may also shelter a resident population. At the population scale, such findings may impact ICES stock assessment and the resulting decisions from EU managers. At the local scale, conservation action could be taken by MPA managers. Besides, this study demonstrates the high potential of archival tags for investigating multi-year behavioural patterns such as site fidelity to offshore spawning areas.},
+    issn = {1054-3139},
+    doi = {10.1093/icesjms/fsy086},
+    url = {https://doi.org/10.1093/icesjms/fsy086},
+    eprint = {https://academic.oup.com/icesjms/article-pdf/76/2/501/31238510/fsy086.pdf},
 }
 
 @article{pontual_seabass_migration_2023,
@@ -33,7 +38,7 @@
     number = {5},
     pages = {1041-1066},
     keywords = {archival tag, Atlantic halibut, geolocation accuracy and precision, Gulf of St. Lawrence, Hidden Markov Model, pop-up satellite archival tag},
-    doi = {https://doi.org/10.1111/faf.12568},
+    doi = {10.1111/faf.12568},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12568},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/faf.12568},
[…]
