### 2025-03-25 fcf3672
reverted previous changes + removed useless classes in healpy.py.
