### 2025-03-26 9b2004a
added version number after any computation in helpers.
