### 2025-03-27 6bbf10f
moved the addition of package versioning information in a function + added 'field_info' key in diff.
