+++ pangeo_fish/helpers.py
@@ -3,6 +3,7 @@ import os
 import re
 import sys
 import warnings
+from importlib.metadata import version
 from pathlib import Path
 
 import fsspec
@@ -25,7 +26,6 @@ from toolz.dicttoolz import valfilter
 from toolz.functoolz import curry  # to change
 from xhealpixify import HealpyGridInfo, HealpyRegridder
 
-import pangeo_fish
 import pangeo_fish.distributions as distrib
 from pangeo_fish.acoustic import emission_probability
 from pangeo_fish.cf import bounds_to_bins
@@ -69,6 +69,14 @@ __all__ = [
     "render_frames",
     "render_distributions",
 ]
+
+
+def _get_package_versions():
+    return {"package_versions": {k: version(k) for k in ["pangeo-fish", "xhealpixify"]}}
+
+
+def _add_package_versions(ds: xr.Dataset):
+    return ds.assign_attrs(ds.attrs | _get_package_versions())
 
 
 def _plot_in_figure(ds: xr.Dataset, **plot_kwargs) -> Figure:
@@ -456,10 +464,14 @@ def compute_diff(
         ),
         other_dim="obs",
     ).chunk({"time": chunk_time})
-    attrs = tag_log.attrs | {
-        "relative_depth_threshold": relative_depth_threshold,
-        "version_number": pangeo_fish.__version__,
-    }
+    attrs = (
+        tag_log.attrs
+        | _get_package_versions()
+        | {
+            "relative_depth_threshold": relative_depth_threshold,
+            "field_info": reference_model.attrs,
+        }
+    )
     diff = (
         diff_z(reference_model, reshaped_tag, depth_threshold=relative_depth_threshold)
         .assign_attrs(attrs)
@@ -585,9 +597,7 @@ def regrid_dataset(
 
     # adds the attributes found in `ds` as well as `min_vertices`
     attrs = ds.attrs.copy()
-    attrs.update(
-        {"min_vertices": min_vertices, "version_number": pangeo_fish.__version__}
-    )
+    attrs.update(_get_package_versions() | {"min_vertices": min_vertices})
     reshaped = reshaped.assign_attrs(attrs)
 
     if save:
@@ -718,14 +728,8 @@ def compute_emission_pdf(
             )
         )
     )  # type: xr.Dataset
-    attrs = diff_ds.attrs.copy()
-    attrs.update(
-        {
-            "differences_std": differences_std,
-            "recapture_std": recapture_std,
-            "version_number": pangeo_fish.__version__,
-        }
-    )
+    attrs = diff_ds.attrs.copy() | _get_package_versions()
+    attrs.update({"differences_std": differences_std, "recapture_std": recapture_std})
     emission_pdf = emission_pdf.assign_attrs(attrs)
     emission_pdf = emission_pdf.chunk({"time": chunk_time} | {d: -1 for d in dims})
 
@@ -812,11 +816,10 @@ def compute_acoustic_pdf(
         chunk_time=chunk_time,
         dims=dims,
     )
-    attrs = emission_ds.attrs.copy()
+    attrs = emission_ds.attrs.copy() | _get_package_versions()
     attrs.update(
         {
             "receiver_buffer": str(receiver_buffer),
-            "version_number": pangeo_fish.__version__,
         }
     )
     acoustic_pdf = acoustic_pdf.assign_attrs(attrs)
@@ -913,7 +916,7 @@ def combine_pdfs(
         combined.attrs.update(ds.attrs)
         if combined.coords.get("cell_ids", None) is not None:
             combined["cell_ids"].attrs.update(ds["cell_ids"].attrs)
-    ds.attrs.update({"version_number": pangeo_fish.__version__})
+    ds.attrs.update(_get_package_versions())
 
[…]
