### 2025-03-28 731f510
revised metadata management in helpers.
