### 2025-03-28 b084b10
Improve metadata transfer across the calculations. (#184)
* updated CI's requirements.

* added version number after any computation in helpers.

* moved the addition of package versioning information in a function + added 'field_info' key in diff.

* revised metadata management in helpers.
