### 2025-03-30 1b1d439
Update CITATION.cff (#185)
