+++ CITATION.cff
@@ -19,7 +19,6 @@ authors:
   - family-names: "Woillez"
     given-names: "Mathieu"
     orcid: "https://orcid.org/0000-0002-1032-2105"
-    affiliation:
   - family-names: "Fouilloux"
     given-names: "Anne"
     orcid: "https://orcid.org/0000-0002-1784-2920"
