### 2025-03-30 2ccf999
Create CONTRIBUTING.md
