+++ CONTRIBUTING.md
@@ -1,0 +1,25 @@
+# Contribution Guidelines
+
+Thank you for your interest in contributing to **pangeo-fish**!
+
+## How to contribute?
+
+- Fork the repository.
+- Create a feature branch: `git checkout -b feature/my-feature`.
+- Commit your changes clearly and concisely.
+- Submit a Pull Request against the `main` branch.
+
+## Coding Standards
+
+- Follow existing coding conventions.
+- Provide tests and documentation for new features.
+
+## Issue Reporting
+
+Report issues clearly with steps to reproduce the problem.
+
+## Code of Conduct
+
+Be kind and respectful.
+
+Thank you for helping improve **pangeo-fish**!
