### 2025-03-30 c5437fc
Update CITATION.cff
