+++ CITATION.cff
@@ -1,4 +1,5 @@
 cff-version: 1.2.0
+message: "If you use this software, please cite it using these metadata."
 title: pangeo-fish
 abstract: "Pangeo-Fish is an open, community-driven project designed to leverage the power of the Pangeo ecosystem specifically for fish biologging data analysis and related oceanographic research. The project facilitates scalable, reproducible, and collaborative research by integrating biologging data with environmental datasets using cloud-native technologies."
 message: "If you use this software, please cite it as below."
@@ -28,6 +29,9 @@ authors:
   - family-names: "Odaka"
     given-names: "Tina"
     orcid: "https://orcid.org/0000-0002-1500-0156"
+version: 2025.03.0
+date-released: "2025-03-07"
+license: GPL-3.0
 repository-code: "https://github.com/pangeo-fish/pangeo-fish"
 url: "https://pangeo-fish.readthedocs.io"
 keywords:
@@ -37,6 +41,6 @@ keywords:
   - Xarray
   - Zarr
   - Dask
-license: GPL-3.0
-version: 2025.03.0
-date-released: "2025-03-07"
+identifiers:
+  - type: doi
+    value: 10.5281/zenodo.15110143
