### 2025-03-30 259ce56
Update CITATION.cff
