### 2025-03-30 84f0be6
Update CITATION.cff
