### 2025-03-30 4ffc227
Update CITATION.cff
