+++ CITATION.cff
@@ -29,7 +29,7 @@ authors:
     given-names: "Tina"
     orcid: "https://orcid.org/0000-0002-1500-0156"
 version: 2025.03.3
-date-released: "2025-03-07"
+date-released: "2025-03-31"
 license: GPL-3.0
 repository-code: "https://github.com/pangeo-fish/pangeo-fish"
 url: "https://pangeo-fish.readthedocs.io"
