### 2025-03-30 75a019b
Update README.md (#186)
* Update README.md

* Add badge

* Update README.md

* Create CONTRIBUTING.md

* Update CITATION.cff
