+++ publication/paper.md
@@ -51,24 +51,34 @@ Geo-referenced data plays an important role in understanding and conserving natu
 Biologging, the practice of attaching small devices (called _tags_) to animals for recording behavior, physiology, and environmental data, proves to be invaluable in this field.
 
 As fish can not be tracked directly using tracking devices such as GPS receivers, geolocation models have emerged to estimate fish positions by correlating individual series of physical measurements — e.g. temperature and pressure records — with geophysical reference fields — oceanic temperature and bathymetry — derived from satellite observations and hydrodynamical model outputs.
-The quality of geophysical reference fields is therefore of paramount importance, which depends on the spatio-temporal resolution at which the physical processes are considered and modeled.
-However, higher resolutions involves more data, that requires significant computing power and storage capacity.
 
-To address this challenge, we developed a Python package for fish tracking estimation, named **pangeo-fish**.
-It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source tools for interactive data analysis in the field of big data marine and geoscience.
+<!-- The quality of geophysical reference fields is therefore of paramount importance, which depends on the spatio-temporal resolution at which the physical processes are considered and modeled.
+However, higher resolutions involves more data, that requires significant computing power and storage capacity. -->
 
-**pangeo-fish** is dedicated to ecologists to process recovered archival tag data and fish tracking estimation.
-These fish geolocation models are critical for better understanding fish movements and behaviors, and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
+Beside the difficulty of working with vast earth science datasets (due to their size and diversity), there is no open source implementation for biologged fish tracking.
+Yet, these fish geolocation models are critical for better understanding fish behavior and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
+
+To address this challenge, we developed a Python package, named **pangeo-fish**, for <u>fish tracking estimation</u>.
+It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source environment for interactive data analysis in the field of big data marine and geoscience.
+
+<!-- **pangeo-fish** is dedicated to ecologists to process recovered archival tag data and <u>fish tracking estimation</u>. -->
+<!-- These fish geolocation models are critical for better understanding fish movements and behaviors, and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation. -->
 
 # Statement of need
 
 Biologging consists in attaching onto (or sometimes inserting into) an animal an
 electronic device that will record in its memory physical and/or geochemical parameters as a function of time so that scientists can reconstruct the activity of the animal, the characteristics of the environment it travels in and the interactions between the two.
-These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, including diving and activity patterns, habitat selection, swimming behavior, mating behavior, energy use and interaction with environment.
+
+<!-- These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, including diving and activity patterns, habitat selection, swimming behavior, mating behavior, energy use and interaction with environment. -->
+
+These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, such as diving and activity patterns, energy use and interaction with environment.
 
 ![Promotion of the FISH-INTEL tagging campaign.\label{fig:fishintel}](fishintel.png){ width=35% }
 
[…]
