### 2025-03-31 2434f16
Paper revision (#174)
* fixed citation.

* attempted to better use the figures.

* fixed DST typo.

* shortened the paper.
