+++ publication/paper.md
@@ -50,49 +50,78 @@ bibliography: paper.bib
 Geo-referenced data plays an important role in understanding and conserving natural resources, particularly when investigating biological phenomena such as fish migration and habitat uses.
 Biologging, the practice of attaching small devices (called _tags_) to animals for recording behavior, physiology, and environmental data, proves to be invaluable in this field.
 
-As fish can not be tracked directly using tracking devices such as GPS receivers, geolocation models have emerged to estimate fish positions by correlating individual series of physical measurements — e.g. temperature and pressure records — with geophysical reference fields - oceanic temperature and bathymetry - derived from satellite observations and hydrodynamical model outputs. The quality of geophysical reference fields is therefore of paramount importance, which depends on the spatio-temporal resolution at which the physical processes are considered and modeled.
-However, higher resolutions involves more data, that requires significant computing power and storage capacity.
+As fish can not be tracked directly using tracking devices such as GPS receivers, geolocation models have emerged to estimate fish positions by correlating individual series of physical measurements — e.g. temperature and pressure records — with geophysical reference fields — oceanic temperature and bathymetry — derived from satellite observations and hydrodynamical model outputs.
 
-To address this challenge, we developed a Python package for fish tracking estimation, named **pangeo-fish**.
-It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source tools for interactive data analysis in the field of big data marine and geoscience.
+<!-- The quality of geophysical reference fields is therefore of paramount importance, which depends on the spatio-temporal resolution at which the physical processes are considered and modeled.
+However, higher resolutions involves more data, that requires significant computing power and storage capacity. -->
 
-**pangeo-fish** is dedicated to ecologists to process recovered archival tag data and fish tracking estimation.
-These fish geolocation models are critical for better understanding fish movements and behaviors, and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
+Beside the difficulty of working with vast earth science datasets (due to their size and diversity), there is no open source implementation for biologged fish tracking.
+Yet, these fish geolocation models are critical for better understanding fish behavior and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
+
+To address this challenge, we developed a Python package, named **pangeo-fish**, for <u>fish tracking estimation</u>.
+It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source environment for interactive data analysis in the field of big data marine and geoscience.
+
+<!-- **pangeo-fish** is dedicated to ecologists to process recovered archival tag data and <u>fish tracking estimation</u>. -->
+<!-- These fish geolocation models are critical for better understanding fish movements and behaviors, and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation. -->
 
 # Statement of need
 
-Bio-logging consists in attaching onto (or sometimes inserting into) an animal an
+Biologging consists in attaching onto (or sometimes inserting into) an animal an
 electronic device that will record in its memory physical and/or geochemical parameters as a function of time so that scientists can reconstruct the activity of the animal, the characteristics of the environment it travels in and the interactions between the two.
-These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, including diving and activity patterns, habitat selection, swimming behavior, mating behavior, energy use, interaction with environment.
-However, unlike terrestrial animals, marine mammals or large sharks (which regularly return to the surface), whose positions can be directly estimated using ARGOS or GPS technologies, tracking fish underwater is challenging.
+
[…]
