+++ publication/paper.md
@@ -52,31 +52,20 @@ Biologging, the practice of attaching small devices (called _tags_) to animals f
 
 As fish can not be tracked directly using tracking devices such as GPS receivers, geolocation models have emerged to estimate fish positions by correlating individual series of physical measurements — e.g. temperature and pressure records — with geophysical reference fields — oceanic temperature and bathymetry — derived from satellite observations and hydrodynamical model outputs.
 
-<!-- The quality of geophysical reference fields is therefore of paramount importance, which depends on the spatio-temporal resolution at which the physical processes are considered and modeled.
-However, higher resolutions involves more data, that requires significant computing power and storage capacity. -->
-
 Beside the difficulty of working with vast earth science datasets (due to their size and diversity), there is no open source implementation for biologged fish tracking.
 Yet, these fish geolocation models are critical for better understanding fish behavior and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
 
 To address this challenge, we developed a Python package, named **pangeo-fish**, for <u>fish tracking estimation</u>.
 It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source environment for interactive data analysis in the field of big data marine and geoscience.
 
-<!-- **pangeo-fish** is dedicated to ecologists to process recovered archival tag data and <u>fish tracking estimation</u>. -->
-<!-- These fish geolocation models are critical for better understanding fish movements and behaviors, and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation. -->
-
 # Statement of need
 
 Biologging consists in attaching onto (or sometimes inserting into) an animal an
 electronic device that will record in its memory physical and/or geochemical parameters as a function of time so that scientists can reconstruct the activity of the animal, the characteristics of the environment it travels in and the interactions between the two.
 
-<!-- These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, including diving and activity patterns, habitat selection, swimming behavior, mating behavior, energy use and interaction with environment. -->
-
 These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, such as diving and activity patterns, energy use and interaction with environment.
 
 ![Promotion of the FISH-INTEL tagging campaign.\label{fig:fishintel}](fishintel.png){ width=35% }
-
-<!-- However, unlike terrestrial animals, marine mammals or large sharks (which regularly return to the surface), whose positions can be directly estimated using ARGOS or GPS technologies, tracking fish underwater is challenging.
-To address this issue, various tagging experiments have been conducted on a variety of fish species [@spanish_tagging; @skagerrak_tagging], and methods have been proposed for approximating the fish locations, referred to as geolocation models [@pontual_seabass_migration_2023; @woillez_hmm-based_2016]. -->
 
 However, unlike terrestrial animals or marine mammals, whose positions can be directly estimated using ARGOS or GPS technologies, tracking fish underwater is challenging.
 To address this issue, various tagging experiments have been conducted on a variety of fish species [@spanish_tagging; @skagerrak_tagging], and methods have been proposed for approximating the fish locations, referred to as geolocation models [@pontual_seabass_migration_2023; @woillez_hmm-based_2016].
@@ -95,11 +84,6 @@ The data from archival tags can offer detailed insights into vertical movement p
 
 \autoref{fig:tag} shows an example of an acoustic tag as well as a DST.
 
-<!-- Over the past two decades, several geolocation methods have been developed to track fish movements [@metcalfe1997tracking; @hunter2003geolocation; @royer2005state; @patterson2017statistical; @pedersen2008geolocation; @woillez_hmm-based_2016].
[…]
