### 2025-03-31 #7938 @QuentinMaz
Hi @kthyng,

I have fixed the DOI issues and shortened the paper. In the meantime, I figured out that the editorial bot seems to account for Markdown comments, so I removed and got (see above) 1870 words.

I have just dumped the actual content of the paper to count the words and I got around 1600 words (including captions and code blocks, excluding references and the headers).

Should we reduce the paper or is it short enough?

Thanks!
Quentin
