### 2025-04-01 a189240
added dependencies for the JupyterLab (binder).
