+++ .binder/environment.yml
@@ -1,8 +1,42 @@
 channels:
   - conda-forge
 dependencies:
+  # development
   - python=3.11
   - jupyterlab
+  # I/O
+  - fsspec
+  - h5netcdf
+  - intake-xarray
+  - zstandard
+  # units, bounds, and cf-conventions
+  - cf_xarray>=0.10.4
+  - pint-xarray
+  # general computing
+  - flox
+  - xarray
+  - more-itertools
+  - cytoolz
+  - toolz
+  - scipy
+  - opt_einsum
+  - sparse
+  # geographic manipulations
+  - movingpandas
+  - geopandas
+  - cartopy
+  - healpy
+  - scikit-learn
+  - astropy
+  # video processing
+  - ffmpeg
+  # distributed computing
+  - dask
+  # code formatting
+  - rich
+  - rich-click
   - pip
   - pip:
+      - cdshealpix
+      - dask-hpcconfig
       - pangeo-fish
