### 2025-04-01 c8558ee
Add badges (#188)
* updated CI's requirements.

* added license, last commit and binder badges.

* added dependencies for the JupyterLab (binder).

* fixed binder's dependencies.

* used Dockerfile for binder.

* WIP: moving requirement files...

* WIP: trying again with .yml specification.
