### 2025-04-01 c5bcb78
added helpers.normalize_pdf() + tip in notebook.
