+++ pangeo_fish/helpers.py
@@ -61,6 +61,7 @@ __all__ = [
     "compute_emission_pdf",
     "compute_acoustic_pdf",
     "combine_pdfs",
+    "normalize_pdf",
     "optimize_pdf",
     "predict_positions",
     "plot_trajectories",
