### 2025-04-02 b24859f
Addition of a high-level pdf normalization function (#193)
* updated CI's requirements.

* added helpers.normalize_pdf() + tip in notebook.

* fixed typo in notebook.

* updated documentation.
