### 2025-04-02 bee1ad8
Fix: [adding initial probability as a parameter] (refs #192)
