### 2025-04-07 #7938 @kthyng
@QuentinMaz I think a shorter paper is better and I would suggest moving most of the "Key Features" section to your docs so you can keep the paper high level and concise.

@RichardLitt This submission is ready to move forward, though perhaps with some changes to the paper in the meantime.
