### 2025-04-11 66f3886
added initial_std in metadata & changed initial_std value in the notebook.
