### 2025-04-14 fb9a7be
updated filtering functions with predictor indices.
