### 2025-04-14 f37bf97
updated search classes with predictor indices.
