+++ pangeo_fish/hmm/optimize/scipy.py
@@ -1,6 +1,8 @@
 import itertools
+import warnings
 
 import more_itertools
+import numpy as np
 import scipy.optimize
 import xarray as xr
 
@@ -71,18 +73,17 @@ class GridSearch:
 
 class EagerBoundsSearch:
     """
-    Optimize estimator parameters within an interval
+    Class for optimizing the parameters of an Estimator within an interval.
 
     Parameters
     ----------
     estimator : Estimator
         The estimator object. Has to have the `set_params(**params) -> Estimator` and
-        `score(data) -> float` methods. Only a single parameter is supported at the
-        moment.
+        `score(data) -> float` methods. Each parameter is git optimized individually.
     param_bounds : sequence of float
-        A sequence containing lower and upper bounds for the parameter.
+        A sequence containing lower and upper bounds for the parameters.
     optimizer_kwargs : mapping, optional
-        Additional parameters for the optimizer
+        Additional parameters for the optimizer.
     """
 
     def __init__(self, estimator, param_bounds, *, optimizer_kwargs={}):
@@ -90,7 +91,147 @@ class EagerBoundsSearch:
         self.param_bounds = tuple(float(v) for v in param_bounds)
         self.optimizer_kwargs = optimizer_kwargs
 
-    def fit(self, X):
+        lower, upper = self.param_bounds
+
+        if lower >= upper:
+            raise ValueError(
+                "The lower bound must be strictly lower than the upper one."
+            )
+
+    def fit_single_parameter(self, X: xr.Dataset):
+        """Optimize the score of the estimator with a single parameter.
+
+        Parameters
+        ----------
+        X : xarray.Dataset
+            The input data.
+
+        Returns
+        -------
+        estimator
+            The estimator with optimized parameters.
+        """
+
+        def f(sigma, X):
+            # computing is important to avoid recomputing as many times as the result is used
+            result = self.estimator.set_params(sigmas=[sigma]).score(X)
+            if not hasattr(result, "compute"):
+                return float(result)
+
+            return float(result.compute())
+
+        if "predictor_index" in X:
+            if X["predictor_index"].dtype != np.int32:
+                X["predictor_index"] = X["predictor_index"].astype(np.int32)
+        else:
+            X = X.assign(
+                predictor_index=("time", np.zeros(X["time"].size).astype(np.int32))
+            )
+
+        lower, upper = self.param_bounds
+        result = scipy.optimize.fminbound(
+            f, lower, upper, args=(X,), **self.optimizer_kwargs
+        )
+
+        return self.estimator.set_params(sigmas=[result.item()])
+
+    def fit_multivariate_parameter(self, X: xr.Dataset):
+        """
+        .. warnings::
+            Not implemented yet.
+        """
+        raise NotImplementedError()
+
+    def fit_multiple_single_parameters(
+        self, X: xr.Dataset, group_name: str = "predictor_index"
+    ):  # , time_slice_indices: list[int | None]):
+        """Optimize the score of the estimator.
+
+        .. warning::
+            Each parameter ``sigma`` is individually optimized with a subset of ``X`` defined by the entry ``X.groupby(group_name)``.
+            As such, it is **not** a single, multivariate search.
+
+        Parameters
+        ----------
+        X : xarray.Dataset
+            The input data.
+        group_name : str, default: "predictor_index"
+            Name of the variable for groupying ``X``. For each group, a single-parameter Estimator is optimized.
+
+        Returns
+        -------
+        estimator
+            The unified estimator with all the optimized parameters found for the groups.
+
+        See Also
+        --------
+        EagerBoundsSearch.fit_multivariate_parameter
+        """
+
+        estimators = []
+
[…]
