### 2025-04-14 b83d8de
updated helper functions and notebook: untested yet.
