+++ pangeo_fish/helpers.py
@@ -1,4 +1,5 @@
 import inspect
+import json
 import os
 import re
 import sys
@@ -15,7 +16,6 @@ import matplotlib.pyplot as plt
 # import hvplot.xarray
 import movingpandas  # noqa: F401
 import numpy as np
-import pandas as pd
 import pint
 import s3fs
 import tqdm
@@ -1095,6 +1095,9 @@ def optimize_pdf(
 ) -> dict:
     """Optimize a temporal probability distribution.
 
+    .. note::
+        To minimize multiple sigma parameters, define the predictors' indices in ``ds["predictor_index"]``.
+
     Parameters
     ----------
     ds : xarray.Dataset
@@ -1123,8 +1126,12 @@ def optimize_pdf(
     Returns
     -------
     params : dict
-        A dictionary containing the optimization results (mainly, the sigma value of the Brownian movement model)
+        A dictionary containing the optimization results (mainly, the sigma value of the Brownian movement model).
+    emission : xr.Dataset
+        The original input dataset with additional attributes. In case of multi-sigma optimization, it also adds the variable "sigma".
     """
+
+    predictor_index = "predictor_index"
 
     # it is important to compute before re-indexing? Yes.
     ds = ds.compute()
@@ -1140,18 +1147,46 @@ def optimize_pdf(
     )
     predictor_factory = _get_predictor_factory(ds=ds, truncate=truncate, dims=dims)
 
-    estimator = EagerEstimator(sigma=None, predictor_factory=predictor_factory)
-    ds.attrs["max_sigma"] = max_sigma  # limitation of the helper
+    estimator = EagerEstimator(sigmas=[None], predictor_factory=predictor_factory)
 
     optimizer = EagerBoundsSearch(
         estimator,
         (1e-4, ds.attrs["max_sigma"]),
         optimizer_kwargs={"disp": 3, "xtol": tolerance},
     )
-    optimized = optimizer.fit(ds)
+    optimized = (
+        optimizer.fit_multiple_single_parameters(ds, predictor_index)
+        if predictor_index in ds
+        else optimizer.fit_single_parameter(ds)
+    )
     params = optimized.to_dict()  # type: dict
     params = _update_params_dict(factory=predictor_factory, params=params)
     params.update(_get_package_versions())
+
+    ds = ds.assign_attrs(
+        ds.attrs | {"max_sigma": max_sigma, "sigmas": params["sigmas"]}
+    )
+
+    # adds sigma time indices in `params` and stamps `ds` with a variable `sigma` in case of multi-sigma
+    def _compute_sigma_var(indices: list[list[int]], values: list[float]):
+        # TODO: input checking...
+        var_size = sum([len(sub_list) for sub_list in indices])
+        var_list = [None for _ in range(var_size)]
+        for i, v in enumerate(values):
+            for index in indices[i]:
+                var_list[index] = v
+        assert all([s is not None for s in var_list])
+        return np.array(var_list)
+
+    if "predictor_index" in ds:
+        indices = [
+            list(group_indices)
+            for group_indices in ds.groupby(predictor_index).groups.values()
+        ]
+        sigma_var = _compute_sigma_var(indices, params["sigmas"])
+
+        params["sigma_indices"] = indices
+        ds = ds.assign(sigma=("time", sigma_var))
 
     if save_parameters:
         try:
@@ -1161,15 +1196,20 @@ def optimize_pdf(
                 str_path_to_json = str(path_to_json)
             else:
                 str_path_to_json = _s3_path_to_str(path_to_json)
-            pd.DataFrame.from_dict(params, orient="index").to_json(
-                str_path_to_json, storage_options=storage_options
-            )
+
+            with fsspec.open(
+                str_path_to_json,
+                "w",
[…]
