### 2025-04-15 d90336b
[pre-commit.ci] pre-commit autoupdate (#182)
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.11.0 → v0.11.5](https://github.com/astral-sh/ruff-pre-commit/compare/v0.11.0...v0.11.5)
- [github.com/abravalheri/validate-pyproject: v0.24 → v0.24.1](https://github.com/abravalheri/validate-pyproject/compare/v0.24...v0.24.1)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
