### 2025-04-16 439ccfd
fixed helper functions + added sigma in states.
