+++ pangeo_fish/helpers.py
@@ -1079,6 +1079,18 @@ def _get_max_sigma(
     return max_sigma.item()
 
 
+def _compute_sigma_var(indices: list[list[int]], values: list[float]):
+    """Return a numpy array composed of the values in ``values`` according to the indices in ``indices``."""
+    # TODO: input checking...
+    var_size = sum([len(sub_list) for sub_list in indices])
+    var_list = [None for _ in range(var_size)]
+    for i, v in enumerate(values):
+        for index in indices[i]:
+            var_list[index] = v
+    assert all([s is not None for s in var_list])
+    return np.array(var_list)
+
+
 def optimize_pdf(
     *,
     ds: xr.Dataset,
@@ -1128,7 +1140,7 @@ def optimize_pdf(
     params : dict
         A dictionary containing the optimization results (mainly, the sigma value of the Brownian movement model).
     emission : xr.Dataset
-        The original input dataset with additional attributes. In case of multi-sigma optimization, it also adds the variable "sigma".
+        The original input dataset with additional metadata. In case of multi-sigma optimization, it also adds the variable "sigma".
     """
 
     predictor_index = "predictor_index"
@@ -1147,11 +1159,11 @@ def optimize_pdf(
     )
     predictor_factory = _get_predictor_factory(ds=ds, truncate=truncate, dims=dims)
 
-    estimator = EagerEstimator(sigmas=[None], predictor_factory=predictor_factory)
+    estimator = EagerEstimator(sigmas=None, predictor_factory=predictor_factory)
 
     optimizer = EagerBoundsSearch(
         estimator,
-        (1e-4, ds.attrs["max_sigma"]),
+        (1e-4, max_sigma),
         optimizer_kwargs={"disp": 3, "xtol": tolerance},
     )
     optimized = (
@@ -1168,24 +1180,14 @@ def optimize_pdf(
     )
 
     # adds sigma time indices in `params` and stamps `ds` with a variable `sigma` in case of multi-sigma
-    def _compute_sigma_var(indices: list[list[int]], values: list[float]):
-        # TODO: input checking...
-        var_size = sum([len(sub_list) for sub_list in indices])
-        var_list = [None for _ in range(var_size)]
-        for i, v in enumerate(values):
-            for index in indices[i]:
-                var_list[index] = v
-        assert all([s is not None for s in var_list])
-        return np.array(var_list)
-
     if "predictor_index" in ds:
-        indices = [
+        sigma_indices = [
             list(group_indices)
             for group_indices in ds.groupby(predictor_index).groups.values()
         ]
-        sigma_var = _compute_sigma_var(indices, params["sigmas"])
+        sigma_var = _compute_sigma_var(sigma_indices, params["sigmas"])
 
-        params["sigma_indices"] = indices
+        params["sigma_indices"] = sigma_indices
         ds = ds.assign(sigma=("time", sigma_var))
 
     if save_parameters:
@@ -1249,7 +1251,7 @@ def predict_positions(
     Returns
     -------
     states : xarray.Dataset
-        A geolocation model, i.e., positional temporal probabilities
+        The positional temporal probabilities. In case of multi-sigma optimization, it also adds the variable "sigma".
     trajectories : movingpandas.TrajectoryCollection
         The tracks decoded from `states`
 
@@ -1280,11 +1282,61 @@ def predict_positions(
     ) as file:
         params = json.load(file)
 
-    # checks that the `emission` has the predictors' indices in case of multiple sigma values
-    if (len(params["sigmas"]) > 1) and (predictor_index not in emission):
-        raise ValueError(
-            'The time indices for each sigma value is not defined in the `emission` (entry "predictor_index" is missing).'
[…]
