### 2025-04-16 2bf8dd8
fixed type-like issue in helper.py.
