### 2025-04-16 90591f6
fixed the docstrings in EagerBoundsSearch.fit_multivariate_parameter().
