### 2025-04-16 b3c6937
fixed unit tests of Estimator and EagerBoundsSearch.
