### 2025-04-16 16d29bf
removed the key features section (cf review process).
