+++ publication/paper.md
@@ -116,92 +116,6 @@ The transition matrix between the hidden states is modelled by a Brownian motion
 As such, fitting the geolocation model for a tag's records aims to determine the value of $\sigma$ that maximizes the likelihood of the state sequence (i.e., the fish's trajectory) given the observations.
 The optimal likelihood value reflects the level of residual inconsistency between the tag observed (recorded) and reference data.
 
-# Key Features of pangeo-fish
-
-**pangeo-fish** is a Python software that handles the entire pipeline for reconstructing fish trajectories given sensor records and a reference data, and visualizing the results.
-The key features of **pangeo-fish** are its simple API that follows each stage of the pipeline, and its scalability, which includes parallel computation and streamlined remote data fetching.
-
-## Pre-processing
-
-The framework starts by loading the archival records of a tagged fish:
-
-```python
-from pangeo_fish.helpers import load_tag
-
-tag, tag_log, time_slice = load_tag(tag_root, tag_name, storage_options)
-```
-
-Then, the user selects the reference model (depending on, for example, the studied area or the resolution of the model):
-
-```python
-from pangeo_fish.helpers import load_model
-
-reference_model = load_model(uri, tag_log, time_slice, ...)
-```
-
-In the previous instruction, the oceanic model is reduced to fit the scope of the tag (in time and space) `load_model()`.
-
-The next step involves complex operations, which are however made easily accessible for the user in a few instructions.
-The operations consists of computing the difference between the archival tag and reference data, reshaping the consequent result (to avoid spatial distortions) to finally compute the emission probabilities given the initial (and optionally the final) position(s) of the fish and the reshaped dataset:
-
-```python
-from pangeo_fish.helpers import compute_diff, regrid_dataset, compute_emission_pdf
-
-diff = compute_diff(reference_model, tag_log, ...)[0]
-reshaped = regrid_dataset(diff, ...)[0]
-emission_pdf = compute_emission_pdf(differences, tag["tagging_events"].ds, ...)[0]
-```
-
-## Accounting for acoustic detections
-
-Before normalizing the emission probabilities, the user can include another distribution, based on the possible fish detections by acoustic receivers:
-
-```python
-from pangeo_fish.helpers import compute_acoustic_pdf, combine_pdfs
-
-acoustic_pdf = compute_acoustic_pdf(emission_pdf, tag, receiver_buffer, ...)
-combined_pdf = combine_pdfs([emission_pdf, acoustic_pdf], ...)
-```
-
-## Model optimization
-
-Finally, the normalized distributions are fitted to find the optimal $\sigma$, and the fish's positions can be extracted (**pangeo-fish** features several options for reconstructing trajectories):
-
-```python
-from pangeo_fish.helpers import optimize_pdf, predict_positions
-
-parameters = optimize_pdf(combined_pdf, maximum_speed, save_parameters=True, ...)
-states, trajectories = predict_positions(path_to_previous_results, track_modes, ...)
-```
-
-## Result analysis and visualization
-
-Every result of the step described above can be easily visualized for analysis with **pangeo-fish**.
-For instance, the time series of the archival tag's can be easily visualized with:
-
-```python
-from pangeo_fish.helpers import plot_tag
-
-plot = plot_tag(tag, tag_log, save_html=True, ...)
-```
-
-The user can also plot an interactive visualization of the trajectories:
-
-```python
-from pangeo_fish.helpers import plot_trajectories
-
-plot = plot_trajectories(path_to_previous_results, track_modes, save_html=True)
-```
-
-As for the state and emission distributions, a video of their evolution (alongside of each other) can be rendered:
-
-```python
-from pangeo_fish.helpers import open_distributions, render_distributions
-
[…]
