### 2025-04-16 4576b2a
removed stamping of the emission dataset with the variable sigma + updated metadata of emission (and so states) dataset(s).
