### 2025-04-22 #7938 @pnsaevik
Hi @RichardLitt, this sounds like an interesting and useful package. Unfortunately I don't have time to review on this side of summer.
