### 2025-05-05 430f31f
added a helper function for adding parameters' indices based on acoustic detections.
