+++ pangeo_fish/helpers.py
@@ -16,6 +16,7 @@ import matplotlib.pyplot as plt
 # import hvplot.xarray
 import movingpandas  # noqa: F401
 import numpy as np
+import pandas as pd
 import pint
 import s3fs
 import tqdm
@@ -45,7 +46,7 @@ from pangeo_fish.io import (
 )
 from pangeo_fish.pdf import combine_emission_pdf, normal
 from pangeo_fish.tags import adapt_model_time, reshape_by_bins, to_time_slice
-from pangeo_fish.utils import temporal_resolution
+from pangeo_fish.utils import haversine_distance, temporal_resolution
 from pangeo_fish.visualization import filter_by_states, plot_map, render_frame
 
 __all__ = [
@@ -53,6 +54,7 @@ __all__ = [
     "reshape_to_2d",
     "load_tag",
     "update_stations",
+    "compute_detection_time_intervals",
     "plot_tag",
     "load_model",
     "compute_diff",
@@ -62,6 +64,7 @@ __all__ = [
     "compute_acoustic_pdf",
     "combine_pdfs",
     "normalize_pdf",
+    "stamp_parameter_indices",
     "optimize_pdf",
     "predict_positions",
     "plot_trajectories",
@@ -258,6 +261,102 @@ def update_stations(
     else:
         tag["stations"] = ds
     return tag
+
+
+def _find_time_intervals(df: pd.DataFrame, min_time: pd.Timedelta, min_dist: float):
+    def _is_far_enough(
+        lon1: float, lat1: float, lon2: float, lat2: float, min_dist: float
+    ):
+        dist = haversine_distance(lon1, lon2, lat1, lat2)
+        return dist > min_dist
+
+    if not all([k in df.columns for k in ["id", "longitude", "latitude"]]):
+        raise ValueError(
+            'The dataframe must have the columns "id", "longitude" and "latitude".'
+        )
+
+    time_delta = pd.Timedelta(0)
+    first_row = df.iloc[0]
+    last_time = first_row.name  # type: pd.Timestamp
+    last_point = (
+        first_row.longitude,
+        first_row.latitude,
+    )  # type: tuple[np.float64, np.float64]
+
+    times = [last_time]
+
+    for curr_time, row in df.iloc[1:].iterrows():
+        time_delta += curr_time - last_time
+
+        if _is_far_enough(*last_point, row.longitude, row.latitude, min_dist):
+            times.append(last_time)
+            times.append(curr_time)
+            time_delta = pd.Timedelta(0)
+
+            last_point = (row.longitude, row.latitude)
+        if time_delta > min_time:
+            times.append(curr_time)
+            time_delta = pd.Timedelta(0)
+
+            last_point = (row.longitude, row.latitude)
+
+        last_time = curr_time
+    return times
+
+
+def compute_detection_time_intervals(
+    *,
+    tag: xr.DataTree,
+    min_time: pd.Timedelta = pd.Timedelta("4days"),
+    min_dist: float = 17.0,
+    **kwargs,
+):
+    """
+    Compute the time intervals based on the acoustic detections of a fish tag.
+
+    Parameters
+    ----------
+    tag : xarray.DataTree
+        The fish tag, assumed to have an entry "acoustic".
+    min_time : pandas.Timedelta, optional
+        Minimum time length for the intervals (defaults to ``pd.Timedelta("4days")``).
+    min_dist : float, default: 17.0
+        Minimum distance between the detections to set apart different intervals, in kilometers.
+
+    Returns
+    -------
+    array of pandas.Timestamps
+        The list of timestamps that delimit the detections.
+    """
+
+    def _mapper(string: str):
+        keys = ["id", "longitude", "latitude"]
+        for k in keys:
[…]
