### 2025-05-06 6882bc5
added a function for reindexing a dataset.
