### 2025-05-06 023795e
added a notebook that illustrates multi-sigma optimization + updated documentation accordingly.
