+++ docs/notebooks.rst
@@ -1,0 +1,11 @@
+Tutorials
+=========
+
+The following pages introduce different aspects and features of `pangeo-fish`.
+
+.. toctree::
+   :maxdepth: 1
+   :hidden:
+
+   notebook1
+   notebook2
