### 2025-05-07 67ed8b7
improved plotting of the trajectories in the new notebook.
