### 2025-05-21 #7938 @RichardLitt
Thanks!

@rsignell and @maxrjones: Assuming no conflicts of interest, would you be OK reviewing this?
