### 2025-05-21 #7938 @maxrjones
> @rsignell and @maxrjones: Assuming no conflicts of interest, would you be OK reviewing this?

where can I find the conflict of interest policy? I work for Development Seed, which is also funded on the same fish tracking project, though I am not part of that project.
