### 2025-05-21 #7938 @tinaok
Development Seed did not participate in development of pangeo-fish. They were in development of GFTS platform, we've used GFTS platform and run pangeo-fish on it.
