### 2025-05-21 #7938 @maxrjones
> Development Seed did not participate in development of pangeo-fish. They were in development of GFTS platform, we've used GFTS platform and run pangeo-fish on it.

Thanks for clarifying Tina. I am comfortable and willing to review this submission if others are also alright with this.
