### 2025-05-22 #7938 @rsignell
I'm happy to review. No conflict of interest here!
