### 2025-06-02 #8332 @editorialbot
[REVIEW]: pangeo-fish: A Python package for studying fish movement using biologging and earth science data
**Submitting author:** @keewis (<a href="http://orcid.org/0000-0002-4254-8002">Justus Magin</a>)
**Repository:** https://github.com/pangeo-fish/pangeo-fish
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 2026.05.0
