### 2025-06-09 df383a3
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.11.5 → v0.11.13](https://github.com/astral-sh/ruff-pre-commit/compare/v0.11.5...v0.11.13)
