+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
       - id: blackdoc-autoupdate-black
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.5
+    rev: v0.11.13
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
