+++ ci/requirements/environment.yaml
@@ -13,7 +13,7 @@ dependencies:
   - pint-xarray
   # general computing
   - flox
-  - xarray<=2025.04.0
+  - xarray<2025.04.0
   - more-itertools
   - cytoolz
   - toolz
