### 2025-06-13 3b2fe4a
healpix-convol version
