+++ ci/requirements/environment.yaml
@@ -39,7 +39,7 @@ dependencies:
   - pip:
       - cdshealpix
       - xdggs
-      - healpix-convolution
+      - healpix-convolution>=0.1.2
       - dask-hpcconfig
       - imageio[ffmpeg]
       # - git+https://github.com/iaocea/xarray-healpy.git@0ffca6058f4008f4f22f076e2d60787fcf32ac82
