### 2025-06-16 pangeo-fish/pangeo-fish#207 @platipodium
Fix installation documentation by updating path to environment.yml
Please correct documentation to reflect

`conda env create -n pangeo-fish -f environment.yml` with updated path `ci/requirments/enivronment.yaml`

Related https://github.com/openjournals/joss-reviews/issues/8332#issuecomment-2932745058
