### 2025-06-16 #8332 @platipodium
- [ ] Fixed installation docs: https://github.com/pangeo-fish/pangeo-fish/issues/207
- [ ]
