### 2025-06-16 pangeo-fish/pangeo-fish#208 @platipodium
Fix installation docs by helping to set up jupyter
The installation docs and `environment.yaml` do not consider that jupyter itself is required.  Elaborate documentation to support first-time users.

Related: https://github.com/openjournals/joss-reviews/issues/8332#issuecomment-2932745058
