### 2025-06-18 7b21cae
Update helpers.py
from ref_lvl to refinement_level
