### 2025-06-18 2a740cd
Changes healpix convention video (#205)
* new branch

* change for healpix convention

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* correcting CI

* correct CI

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* CI

* Trigger CI

* healpix-convol version

* Update helpers.py

from ref_lvl to refinement_level

* Update pangeo-fish.ipynb

from ref_lvl to refinement_level

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
