### 2025-06-18 3e58011
[pre-commit.ci] pre-commit autoupdate (#198)
updates:
- [github.com/keewis/blackdoc: v0.3.9 → v0.4.0](https://github.com/keewis/blackdoc/compare/v0.3.9...v0.4.0)
- [github.com/astral-sh/ruff-pre-commit: v0.11.5 → v0.11.13](https://github.com/astral-sh/ruff-pre-commit/compare/v0.11.5...v0.11.13)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
