+++ .pre-commit-config.yaml
@@ -17,14 +17,14 @@ repos:
       - id: black-jupyter
 
   - repo: https://github.com/keewis/blackdoc
-    rev: v0.3.9
+    rev: v0.4.0
     hooks:
       - id: blackdoc
         additional_dependencies: ["black==25.1.0"]
       - id: blackdoc-autoupdate-black
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.5
+    rev: v0.11.13
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
