### 2025-06-24 8a2cc97
Modification for reference model compliance
