### 2025-06-25 1f46ebb
allowing Kerchunk reference JSON
