### 2025-06-25 c263ee3
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
