### 2025-06-25 3f6ebe1
Modif for reference model compliance (#211)
* new branch

* change for healpix convention

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* correcting CI

* correct CI

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* CI

* Trigger CI

* healpix-convol version

* Modification for reference model compliance

* allowing Kerchunk reference JSON

* changing notebooks with right ref file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-com
[…]
