### 2025-06-25 3ff3be8
fixing drop_indexes
