### 2025-08-04 87addb6
First commit with adding 'generical' bathy_pdf
