+++ pangeo_fish/bathy.py
@@ -1,0 +1,251 @@
+import numpy as np
+import xarray as xr
+import healpy as hp
+import dask.array as da
+from dask import delayed
+from distributed import LocalCluster
+from tqdm import tqdm
+
+from pangeo_fish.helpers import load_model, compute_diff
+from pangeo_fish.io import open_tag
+from pangeo_fish.cf import bounds_to_bins
+from pangeo_fish.diff import diff_z
+from pangeo_fish.tags import adapt_model_time, reshape_by_bins, to_time_slice
+
+import xdggs
+import matplotlib.pyplot as plt
+
+
+def compute_healpix_histogram_region(
+    ds,
+    nside,
+    nb_depth_bins,
+    chunk_size=500,
+    depth_offset=0
+):
+    """
+    Compute HEALPix histogram only over the geographic region covered by `ds`,
+    using a two-pass approach:
+      1) Identify the unique cells actually touched
+      2) Accumulate histogram on this subset
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+        Contains `elevation` (2D lat×lon) and optionally `stdev`.
+    nside : int
+        HEALPix resolution (power of 2).
+    nb_depth_bins : int
+        Number of bins (excluding the +10 offset).
+    chunk_size : int
+        Number of latitude rows per iteration.
+    depth_offset : float
+        Offset applied to bathymetry before binning.
+    """
+    nest = True
+    bins = nb_depth_bins + 10
+
+    lats = ds.latitude.values
+    lons = ds.longitude.values
+    nlat, nlon = len(lats), len(lons)
+
+    # --- PASS 1: Collect unique used HEALPix cells ---
+    used_cells = set()
+    for i in tqdm(range(0, nlat, chunk_size), desc="Pass 1 – collect cells"):
+        j = min(i + chunk_size, nlat)
+
+        elev = ds.elevation[i:j, :].load().values.flatten()
+        valid = ~np.isnan(elev)
+        if not valid.any():
+            continue
+
+        lat_blk = np.repeat(lats[i:j], nlon)[valid]
+        lon_blk = np.tile(lons, j - i)[valid]
+
+        hidx = hp.ang2pix(nside, lon_blk, lat_blk, lonlat=True, nest=nest)
+        used_cells.update(np.unique(hidx))
+
+    used_cells = np.array(sorted(used_cells), dtype=np.int64)
+    n_cells = used_cells.size
+    cell_to_idx = {cell: idx for idx, cell in enumerate(used_cells)}
+
+    hist = np.zeros((n_cells, bins), dtype=np.float64)
+
+    # --- PASS 2: Accumulate histogram ---
+    for i in tqdm(range(0, nlat, chunk_size), desc="Pass 2 – compute histogram"):
+        j = min(i + chunk_size, nlat)
+
+        elev = ds.elevation[i:j, :].load().values.flatten()
+        st = ds.stdev[i:j, :].load().values.flatten() if "stdev" in ds else np.full_like(elev, 1.0)
+
+        valid = ~np.isnan(elev)
+        if not valid.any():
+            continue
+
+        elev_v = elev[valid]
+        st_v = np.maximum(np.nan_to_num(st[valid], nan=1.0), 0.5)
+        lat_v = np.repeat(lats[i:j], nlon)[valid]
+        lon_v = np.tile(lons, j - i)[valid]
+
+        hidx = hp.ang2pix(nside, lon_v, lat_v, lonlat=True, nest=nest)
[…]
