### 2025-08-05 8cf9c1e
Notebook with generic GEBCO bathy implemented
