### 2025-08-06 #8332 @RichardLitt
@maxrjones Let me know if you're still keen on this.
