### 2025-09-05 75217d6
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
