+++ pangeo_fish/bathy.py
@@ -1,29 +1,26 @@
+import warnings
+
+import dask.array as da
+import healpy as hp
+import matplotlib.pyplot as plt
 import numpy as np
 import xarray as xr
-import healpy as hp
-import dask.array as da
+import xdggs
 from dask import delayed
 from distributed import LocalCluster
 from tqdm import tqdm
 
-from pangeo_fish.helpers import load_model, compute_diff
-from pangeo_fish.io import open_tag
 from pangeo_fish.cf import bounds_to_bins
 from pangeo_fish.diff import diff_z
+from pangeo_fish.helpers import compute_diff, load_model
+from pangeo_fish.io import open_tag
 from pangeo_fish.tags import adapt_model_time, reshape_by_bins, to_time_slice
 
-import xdggs
-import matplotlib.pyplot as plt
-import warnings
-warnings.filterwarnings("ignore", category=RuntimeWarning) 
+warnings.filterwarnings("ignore", category=RuntimeWarning)
 
 
 def compute_healpix_histogram_region(
-    ds,
-    nside,
-    nb_depth_bins,
-    chunk_size=500,
-    depth_offset=0
+    ds, nside, nb_depth_bins, chunk_size=500, depth_offset=0
 ):
     """
     Compute HEALPix histogram only over the geographic region covered by `ds`,
@@ -78,7 +75,11 @@ def compute_healpix_histogram_region(
         j = min(i + chunk_size, nlat)
 
         elev = ds.elevation[i:j, :].load().values.flatten()
-        st = ds.stdev[i:j, :].load().values.flatten() if "stdev" in ds else np.full_like(elev, 1.0)
+        st = (
+            ds.stdev[i:j, :].load().values.flatten()
+            if "stdev" in ds
+            else np.full_like(elev, 1.0)
+        )
 
         valid = ~np.isnan(elev)
         if not valid.any():
@@ -90,7 +91,10 @@ def compute_healpix_histogram_region(
         lon_v = np.tile(lons, j - i)[valid]
 
         hidx = hp.ang2pix(nside, lon_v, lat_v, lonlat=True, nest=nest)
-        depth_idx = np.clip(-elev_v - depth_offset, -10, nb_depth_bins - 0.01).astype(np.int64) + 10
+        depth_idx = (
+            np.clip(-elev_v - depth_offset, -10, nb_depth_bins - 0.01).astype(np.int64)
+            + 10
+        )
 
         mask_zone = np.isin(hidx, used_cells)
         hidx = hidx[mask_zone]
@@ -99,21 +103,21 @@ def compute_healpix_histogram_region(
 
         sum_w = np.zeros_like(st_v)
         for dj in range(-2, 3):
-            sum_w += np.exp(-dj**2 / (2 * st_v**2))
+            sum_w += np.exp(-(dj**2) / (2 * st_v**2))
         inv_sum = 1.0 / sum_w
 
         uniq, inv = np.unique(hidx, return_inverse=True)
         hist_local = np.zeros((uniq.size, bins), dtype=np.float64)
 
         for dj in range(-2, 3):
-            w = np.exp(-dj**2 / (2 * st_v**2)) * inv_sum
+            w = np.exp(-(dj**2) / (2 * st_v**2)) * inv_sum
             idx_shifted = np.clip(depth_idx + dj, 0, bins - 1)
             np.add.at(hist_local, (inv, idx_shifted), w)
 
         for u_idx, cell in enumerate(uniq):
             hist[cell_to_idx[cell], :] += hist_local[u_idx, :]
 
-    with np.errstate(invalid='ignore'):
+    with np.errstate(invalid="ignore"):
[…]
