+++ docs/api.rst
@@ -30,7 +30,7 @@ High-level Functions
    plot_distributions
    render_frames
    render_distributions
-   
+
 
 
 I/O
