### 2025-09-05 67af1bf
all files with bathy.py change with bins selection
