+++ pangeo_fish/bathy.py
@@ -18,6 +18,152 @@ from pangeo_fish.tags import adapt_model_time, reshape_by_bins, to_time_slice
 
 warnings.filterwarnings("ignore", category=RuntimeWarning)
 
+import numpy as np
+import xarray as xr
+from tqdm import tqdm
+import healpy as hp
+
+def compute_healpix_histogram_region_bin_size(
+    ds,
+    nside,
+    nb_depth_bins=None,
+    max_depth_m=None,
+    chunk_size=500,
+    depth_offset=0,
+    depth_bin_size=1
+):
+    """
+    Compute HEALPix histogram over the geographic region covered by `ds`.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+        Contains `elevation` (2D lat×lon) and optionally `stdev`.
+    nside : int
+        HEALPix resolution (power of 2).
+    nb_depth_bins : int or None
+        Number of depth bins (in units of `depth_bin_size`). If provided, used directly.
+    max_depth_m : float or None
+        If provided and nb_depth_bins is None, nb_depth_bins = ceil(max_depth_m / depth_bin_size).
+    chunk_size : int
+        Number of latitude rows per iteration.
+    depth_offset : float
+        Offset applied to bathymetry before binning.
+    depth_bin_size : float
+        Bin width in meters (default=1). Example: 100 → bins de 100 m.
+    """
+
+    # --- determine nb_depth_bins from max_depth_m if needed ---
+    if nb_depth_bins is None:
+        if max_depth_m is None:
+            raise ValueError("Either nb_depth_bins or max_depth_m must be provided.")
+        nb_depth_bins = int(np.ceil(float(max_depth_m) / float(depth_bin_size)))
+    else:
+        nb_depth_bins = int(nb_depth_bins)
+
+    if nb_depth_bins <= 0:
+        raise ValueError("nb_depth_bins must be > 0")
+
+    nest = True
+    # historic behaviour: we keep +10 buffer bins as in original code
+    bins = nb_depth_bins + 10
+
+    lats = ds.latitude.values
+    lons = ds.longitude.values
+    nlat, nlon = len(lats), len(lons)
+
+    # --- PASS 1: Collect unique used HEALPix cells ---
+    used_cells = set()
+    for i in tqdm(range(0, nlat, chunk_size), desc="Pass 1 – collect cells"):
+        j = min(i + chunk_size, nlat)
+
+        elev = ds.elevation[i:j, :].load().values.flatten()
+        valid = ~np.isnan(elev)
+        if not valid.any():
+            continue
+
+        lat_blk = np.repeat(lats[i:j], nlon)[valid]
+        lon_blk = np.tile(lons, j - i)[valid]
+
+        hidx = hp.ang2pix(nside, lon_blk, lat_blk, lonlat=True, nest=nest)
+        used_cells.update(np.unique(hidx))
+
+    used_cells = np.array(sorted(used_cells), dtype=np.int64)
+    n_cells = used_cells.size
+    cell_to_idx = {cell: idx for idx, cell in enumerate(used_cells)}
+
+    hist = np.zeros((n_cells, bins), dtype=np.float64)
+
+    # --- PASS 2: Accumulate histogram ---
+    for i in tqdm(range(0, nlat, chunk_size), desc="Pass 2 – compute histogram"):
+        j = min(i + chunk_size, nlat)
+
+        elev = ds.elevation[i:j, :].load().values.flatten()
+        st = ds.stdev[i:j, :].load().values.flatten() if "stdev" in ds else np.full_like(elev, 1.0)
+
[…]
