### 2025-09-05 032a088
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
