### 2025-09-05 9b082a7
Finalisation modifications bathy.py
