### 2025-09-22 a3dbf2a
Fix README and environment.yaml for first-time users
References:
- https://github.com/pangeo-fish/pangeo-fish/issues/207
- https://github.com/pangeo-fish/pangeo-fish/issues/208
