+++ README.md
@@ -31,12 +31,18 @@ Or set up the development environment:
 ```bash
 git clone https://github.com/pangeo-fish/pangeo-fish.git
 cd pangeo-fish
-conda env create -n pangeo-fish -f environment.yml
+conda env create -n pangeo-fish -f ci/requirements/environment.yaml
 conda activate pangeo-fish
 jupyter lab
 ```
 
-Explore example notebooks located in the [`notebooks`](https://github.com/pangeo-fish/pangeo-fish/tree/main/notebooks) directory.
+### For first-time users
+
+If this is your first time using **Pangeo-Fish**, we recommend setting up the full development environment to ensure all dependencies (including [JupyterLab](https://jupyterlab.readthedocs.io/en/stable/)) are correctly installed.
+Once JupyterLab has started, open the notebook  
+[notebooks/pangeo-fish.ipynb](https://github.com/pangeo-fish/pangeo-fish/blob/main/notebooks/pangeo-fish.ipynb)
+
+to follow the step-by-step tutorial. This is the recommended starting point for new users.
 
 ## 📖 Documentation
 
