### 2025-09-22 0e96e50
adding complement from @rsignell for the the environnment
