### 2025-09-22 8430d4f
adding complement from @rsignell for the the environnment
