### 2025-09-25 1e41642
adding a specific version for kerchunk
