### 2025-09-25 pangeo-fish/pangeo-fish#208 @capetienne
Solved in #218
