### 2025-09-25 #8332 @RichardLitt
@platipodium Take a look at https://github.com/pangeo-fish/pangeo-fish/pull/218 and let me know what you think. Thanks.
