### 2025-09-26 pangeo-fish/pangeo-fish#207 @platipodium
Hmm, this does not seem solved. In Readme, it says

`conda env create -n pangeo-fish -f environment.yml`

But this does not exist, instead there are three more in different paths:

```
08:10:47 ✖  find . -name 'envir*'
./.binder/environment.yml
./ci/requirements/environment.yaml
./docs/environment.yaml
```
