### 2025-09-26 pangeo-fish/pangeo-fish#207 @platipodium
There should not be a separate instruction for  `dask-hpcconfig` as it is already part of the `conda create`
