### 2025-09-26 #8332 @platipodium
> [@platipodium](https://github.com/platipodium) Take a look at [pangeo-fish/pangeo-fish#218](https://github.com/pangeo-fish/pangeo-fish/pull/218) and let me know what you think. Thanks.

Still the documentation of installation is insufficient and contains errors - at least on my platform.  Please reopen https://github.com/pangeo-fish/pangeo-fish/issues/207 @capetienne and fix. I still have not been able to install pangeo-fish with conda, let alone test it.
