### 2025-09-26 pangeo-fish/pangeo-fish#219 @platipodium
Match or explain mismatch in authors for JOSS
Related https://github.com/openjournals/joss-reviews/issues/8332#issuecomment-2932745058

7. The List of authors in the JOSS paper is very different from the git logged author commits.  Please clarify.
8. Please clean up git history with `tinaok` and `Tina Odaka` appearing as different users.
