### 2025-09-26 pangeo-fish/pangeo-fish#220 @platipodium
DRY acknowledgments section in JOSS
Related https://github.com/openjournals/joss-reviews/issues/8332

The Acknowledgement section does not follow DRY principle.  Please make it concise and short (I would avoid a list, but prefer one paragraph).  It is standard practice to not use names, but only acronyms, e.g. T.O. obtained grant XXX.
