### 2025-09-26 pangeo-fish/pangeo-fish#207 @tinaok
Hi @platipodium
thank you for taking care of this.  Could you please indicate what kind of HPC system are you using to test the installation?
thank you, Tina
