### 2025-09-26 pangeo-fish/pangeo-fish#207 @platipodium
No worries, I am happy to test.  Lowering the barrier to adopting a software by proper and working instructions is key, and there are always many pitfalls that are usually only found by testing in your user base.

* I am using a RedHat Linux Cluster with SLURM scheduler.
