### 2025-09-26 b31cdc8
Fix issue #207:Changes in the installation.md + right path +typo fix
