### 2025-09-27 pangeo-fish/pangeo-fish#207 @platipodium
Ok, I see the typos fixed (but need to push to readthedocs).   I would need jupyterlab as a dependency in environment.yml ....
