+++ .github/workflows/pypi.yml
@@ -10,11 +10,11 @@ jobs:
     runs-on: ubuntu-latest
     if: github.repository == 'pangeo-fish/pangeo-fish'
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
         with:
           fetch-depth: 0
 
-      - uses: actions/setup-python@v5
+      - uses: actions/setup-python@v6
         name: Install Python
         with:
           python-version: "3.x"
@@ -46,7 +46,7 @@ jobs:
 
     steps:
       - name: Download the artifacts
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v5
         with:
           name: releases
           path: dist/
