### 2025-10-08 012bfab
DRY principle + paragraph for Acknowledgment
