### 2025-10-28 pangeo-fish/pangeo-fish#219 @tinaok
Thank you for your comment.
> 7. The List of authors in the JOSS paper is very different from the git logged author commits. Please clarify.

For clarity, the main commit handles map to paper authors as follows:
	•	keewis → Justus Magin
	•	QuentinMaz → Quentin Mazouni
	•	capetienne → Etienne Cap
	•	tinaok → Tina Odaka

Other co‑authors (Marine Gonse, Mathieu Woillez, Anne Fouilloux, Jean-Marc Delouis)  contributed substantially through method design, validation with domain data, supervision.

Would you like that I make CRediT-style “Author contributions” here?
