### 2025-10-29 pangeo-fish/pangeo-fish#219 @tinaok
Thanks for the proposition @platipodium.
@capetienne , could you consider adding the full author/contributor list in codemeta.json ? Thank you
