### 2025-10-30 c185b2e
first implementation of Foscat
