### 2025-10-30 6157b01
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
