### 2025-11-04 d486cf6
patch for truncate in predict_pos
