### 2025-11-10 pangeo-fish/pangeo-fish#219 @tinaok
Hi @platipodium ,

>8. Please clean up git history with tinaok and Tina Odaka appearing as different users.

@keewis and @capetienne told me that they are working on it this point some time ago.  I'll check with them the status.

> full author/contributor list (maybe you have already) in your codemeta.json with respective roles assigned.

I'll check the status of this with @capetienne as well.
