### 2025-11-10 #8332 @tinaok
Hi @RichardLitt — thanks for the nudge.

We’re wrapping up the authorship clarification tracked in [pangeo-fish/pangeo-fish#219.
](https://github.com/pangeo-fish/pangeo-fish/issues/219)

I'll coordinate with @capetienne, and get back to you.
