### 2025-11-10 pangeo-fish/pangeo-fish#219 @tinaok
Hi @platipodium ,

> Please clean up git history with tinaok and Tina Odaka appearing as different users.
This should be fixed now with #225
