### 2025-11-10 pangeo-fish/pangeo-fish#226 @platipodium
Clarify copyright
https://github.com/openjournals/joss-reviews/issues/8332#issuecomment-2932745058

I see that the LICENSE is GPL and applies to all your code.  I also searched for COPYRIGHT information and only found this in your LICENSE statement:

```
 If the program does terminal interaction, make it output a short
notice like this when it starts in an interactive mode:

    <program>  Copyright (C) <2024>  <Tina Odaka, Justus Magne, Jean Marc Delouis, Mathiew Woillez, >
    This program comes with ABSOLUTELY NO WARRANTY; for details type `show w'.
    This is free software, and you are welcome to redistribute it
    under certain conditions; type `show c' for details.

The hypothetical commands `show w' and `show c' should show the appropriate
parts of the General Public License.  Of course, your program's commands
might be different; for a GUI interface, you would use an "about box".

  You should also get your employer (if you work as a programmer) or school,
if any, to sign a "copyright disclaimer" for the program, if necessary.
For more information on this, and how to apply and follow the GNU GPL, see
<https://www.gnu.org/licenses/>.
```
