### 2025-11-10 pangeo-fish/pangeo-fish#226 @platipodium
I used reuse software on your project and it found

```
* Files with copyright information: 1 / 122
* Files with license information: 0 / 122
```
