### 2025-11-10 pangeo-fish/pangeo-fish#227 @platipodium
Sort out diverging license statements
https://github.com/openjournals/joss-reviews/issues/8332#issuecomment-2932745058

Your project is licensed under

 * BSD (Readme)
 * GPL (INSTALL)
 * MIT (pyproject.toml)

At least that's what the files in your repo say.  Please either put the entire repository under a single license and be consistent in your meta information file, or use reuse.software to attache different licenses to different files.

It is good practise (read the GPL in your repo) that each and every file contains License/copyright statements; it is not mandatory and you can choose not to do this.  Please state here why you would not choose to do this, or update each file.  I strongly recommend to use reuse.sofware if you choose the latter.
