### 2025-11-19 pangeo-fish/pangeo-fish#229 @platipodium
Running the tutorial fails for me
related testing functionality for https://github.com/openjournals/joss-reviews/issues/8332

I am running the tutorial (thanks for the clear instructions now!) on my Mac M1 and I seem to run into memory problems.  Now that is an old machine with 8 GB memory, but I believe your tutorial should run on non-HPC machines, too.  But maybe the problem is something entirely different:

from block 11 in your tutorial notebook `diff = diff.compute()`

```
2025-11-19 10:09:28,090 - distributed.worker.memory - WARNING - Worker is at 88% memory usage. Pausing worker.  Process memory: 1.76 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:09:28,438 - distributed.worker.memory - WARNING - Worker is at 51% memory usage. Resuming worker. Process memory: 1.03 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:10:09,096 - distributed.worker.memory - WARNING - Worker is at 82% memory usage. Pausing worker.  Process memory: 1.65 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:10:09,124 - distributed.worker.memory - WARNING - Worker is at 47% memory usage. Resuming worker. Process memory: 0.96 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:11:05,121 - distributed.worker.memory - WARNING - Worker is at 88% memory usage. Pausing worker.  Process memory: 1.77 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:11:05,438 - distributed.worker.memory - WARNING - Worker is at 24% memory usage. Resuming worker. Process memory: 496.03 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:12:41,765 - distributed.worker - ERROR - Worker stream died during communication: tcp://127.0.0.1:55186
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 228](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=227), in read
    frames_nosplit = await read_bytes_rw(stream, frames_nosplit_nbytes)
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 367](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=366), in read_bytes_rw
    actual = await stream.read_into(chunk)  # type: ignore[arg-type]
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
tornado.iostream.StreamClosedError: Stream is closed

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2081](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2080), in gather_dep
    response = await get_data_from_worker(
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2887](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2886), in get_data_from_worker
    response = await send_recv(
               ^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py", line 1018](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py#line=1017), in send_recv
    response = await comm.read(deserializers=deserializers)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 237](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=236), in read
    convert_stream_closed_error(self, e)
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 137](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=136), in convert_stream_closed_error
    raise CommClosedError(f"in {obj}: {exc}") from exc
distributed.comm.core.CommClosedError: in <TCP (closed) Ephemeral Worker->Worker for gather local=tcp://127.0.0.1:55194 remote=tcp://127.0.0.1:55186>: Stream is closed
2025-11-19 10:12:41,850 - distributed.worker - ERROR - failed during get data with tcp://127.0.0.1:55166 -> tcp://127.0.0.1:55186
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 962](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=961), in _handle_write
    num_bytes = self.write_to_fd(self._write_buffer.peek(size))
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 1121](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=1120), in write_to_fd
    return self.socket.send(data)  # type: ignore
           ^^^^^^^^^^^^^^^^^^^^^^
BrokenPipeError: [Errno 32] Broken pipe

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 1803](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=1802), in get_data
    response = await comm.read(deserializers=serializers)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 237](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=236), in read
    convert_stream_closed_error(self, e)
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 143](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=142), in convert_stream_closed_error
    raise CommClosedError(f"in {obj}: {exc.__class__.__name__}: {exc}") from exc
distributed.comm.core.CommClosedError: in <TCP (closed)  local=tcp://127.0.0.1:55166 remote=tcp://127.0.0.1:55189>: BrokenPipeError: [Errno 32] Broken pipe
2025-11-19 10:13:27,077 - distributed.worker.memory - WARNING - Worker is at 81% memory usage. Pausing worker.  Process memory: 1.63 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:13:27,245 - distributed.worker.memory - WARNING - Worker is at 27% memory usage. Resuming worker. Process memory: 556.41 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:13:34,761 - distributed.worker.memory - WARNING - Worker is at 82% memory usage. Pausing worker.  Process memory: 1.65 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:13:35,118 - distributed.worker.memory - WARNING - Worker is at 78% memory usage. Resuming worker. Process memory: 1.58 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:13:36,689 - distributed.worker.memory - WARNING - Unmanaged memory use is high. This may indicate a memory leak or the memory may not be released to the OS; see https://distributed.dask.org/en/latest/worker-memory.html#memory-not-released-back-to-the-os for more information. -- Unmanaged memory: 1.28 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:14:08,366 - distributed.worker - ERROR - Worker stream died during communication: tcp://127.0.0.1:55166
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 861](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=860), in _read_to_buffer
    bytes_read = self.read_from_fd(buf)
                 ^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 1113](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=1112), in read_from_fd
    return self.socket.recv_into(buf, len(buf))
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ConnectionResetError: [Errno 54] Connection reset by peer

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2081](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2080), in gather_dep
    response = await get_data_from_worker(
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2887](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2886), in get_data_from_worker
    response = await send_recv(
               ^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py", line 1018](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py#line=1017), in send_recv
    response = await comm.read(deserializers=deserializers)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 237](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=236), in read
    convert_stream_closed_error(self, e)
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 143](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=142), in convert_stream_closed_error
    raise CommClosedError(f"in {obj}: {exc.__class__.__name__}: {exc}") from exc
distributed.comm.core.CommClosedError: in <TCP (closed) Ephemeral Worker->Worker for gather local=tcp://127.0.0.1:55210 remote=tcp://127.0.0.1:55166>: ConnectionResetError: [Errno 54] Connection reset by peer
2025-11-19 10:14:24,390 - distributed.worker.memory - WARNING - Worker is at 93% memory usage. Pausing worker.  Process memory: 1.87 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:14:59,775 - distributed.worker.memory - WARNING - Worker is at 81% memory usage. Pausing worker.  Process memory: 1.63 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:15:00,090 - distributed.worker.memory - WARNING - Worker is at 27% memory usage. Resuming worker. Process memory: 562.30 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:15:15,268 - distributed.worker.memory - WARNING - Worker is at 81% memory usage. Pausing worker.  Process memory: 1.63 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:15:48,702 - distributed.worker.memory - WARNING - Worker is at 83% memory usage. Pausing worker.  Process memory: 1.68 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:15:48,863 - distributed.worker.memory - WARNING - Worker is at 27% memory usage. Resuming worker. Process memory: 570.33 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:15:53,896 - distributed.worker.memory - WARNING - Worker is at 84% memory usage. Pausing worker.  Process memory: 1.68 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:09,036 - distributed.worker.memory - WARNING - Worker is at 81% memory usage. Pausing worker.  Process memory: 1.63 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:09,192 - distributed.worker.memory - WARNING - Worker is at 59% memory usage. Resuming worker. Process memory: 1.19 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:19,102 - distributed.worker.memory - WARNING - Worker is at 80% memory usage. Pausing worker.  Process memory: 1.61 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:19,432 - distributed.worker.memory - WARNING - Worker is at 40% memory usage. Resuming worker. Process memory: 835.52 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:58,260 - distributed.worker.memory - WARNING - Worker is at 82% memory usage. Pausing worker.  Process memory: 1.64 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:58,311 - distributed.worker.memory - WARNING - Worker is at 35% memory usage. Resuming worker. Process memory: 733.94 MiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:58,701 - distributed.worker.memory - WARNING - Worker is at 94% memory usage. Pausing worker.  Process memory: 1.89 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:17:59,074 - distributed.worker.memory - WARNING - Worker is at 46% memory usage. Resuming worker. Process memory: 0.93 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:19:36,000 - distributed.worker.memory - WARNING - Worker is at 89% memory usage. Pausing worker.  Process memory: 1.80 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:19:36,370 - distributed.worker.memory - WARNING - Worker is at 56% memory usage. Resuming worker. Process memory: 1.14 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:21:49,464 - distributed.worker - ERROR - failed during get data with tcp://127.0.0.1:55318 -> tcp://127.0.0.1:55288
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 298](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=297), in write
    raise StreamClosedError()
tornado.iostream.StreamClosedError: Stream is closed

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 1802](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=1801), in get_data
    compressed = await comm.write(msg, serializers=serializers)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 308](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=307), in write
    convert_stream_closed_error(self, e)
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 137](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=136), in convert_stream_closed_error
    raise CommClosedError(f"in {obj}: {exc}") from exc
distributed.comm.core.CommClosedError: in <TCP (closed)  local=tcp://127.0.0.1:55318 remote=tcp://127.0.0.1:55327>: Stream is closed
2025-11-19 10:22:00,361 - distributed.worker.memory - WARNING - Worker is at 91% memory usage. Pausing worker.  Process memory: 1.84 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:22:00,571 - distributed.worker.memory - WARNING - Worker is at 74% memory usage. Resuming worker. Process memory: 1.50 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:22:47,112 - distributed.worker - ERROR - Worker stream died during communication: tcp://127.0.0.1:55318
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 861](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=860), in _read_to_buffer
    bytes_read = self.read_from_fd(buf)
                 ^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 1113](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=1112), in read_from_fd
    return self.socket.recv_into(buf, len(buf))
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ConnectionResetError: [Errno 54] Connection reset by peer

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2081](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2080), in gather_dep
    response = await get_data_from_worker(
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2887](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2886), in get_data_from_worker
    response = await send_recv(
               ^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py", line 1018](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py#line=1017), in send_recv
    response = await comm.read(deserializers=deserializers)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 237](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=236), in read
    convert_stream_closed_error(self, e)
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py", line 143](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/comm/tcp.py#line=142), in convert_stream_closed_error
    raise CommClosedError(f"in {obj}: {exc.__class__.__name__}: {exc}") from exc
distributed.comm.core.CommClosedError: in <TCP (closed) Ephemeral Worker->Worker for gather local=tcp://127.0.0.1:55343 remote=tcp://127.0.0.1:55318>: ConnectionResetError: [Errno 54] Connection reset by peer
2025-11-19 10:23:25,905 - distributed.worker.memory - WARNING - Worker is at 82% memory usage. Pausing worker.  Process memory: 1.64 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:23:26,156 - distributed.worker.memory - WARNING - Worker is at 77% memory usage. Resuming worker. Process memory: 1.55 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:23:51,061 - distributed.worker.memory - WARNING - Worker is at 80% memory usage. Pausing worker.  Process memory: 1.61 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:23:59,444 - distributed.worker.memory - WARNING - Worker is at 82% memory usage. Pausing worker.  Process memory: 1.65 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:23:59,859 - distributed.worker.memory - WARNING - Worker is at 74% memory usage. Resuming worker. Process memory: 1.48 GiB -- Worker memory limit: 2.00 GiB
2025-11-19 10:26:41,311 - distributed.worker - ERROR - Worker stream died during communication: tcp://127.0.0.1:55397
Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 861](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=860), in _read_to_buffer
    bytes_read = self.read_from_fd(buf)
                 ^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py", line 1113](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/tornado/iostream.py#line=1112), in read_from_fd
    return self.socket.recv_into(buf, len(buf))
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ConnectionResetError: [Errno 54] Connection reset by peer

Traceback (most recent call last):
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2081](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2080), in gather_dep
    response = await get_data_from_worker(
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py", line 2887](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/worker.py#line=2886), in get_data_from_worker
    response = await send_recv(
               ^^^^^^^^^^^^^^^^
  File "[/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py", line 1018](http://localhost:8888/opt/conda/envs/pangeo-fish/lib/python3.12/site-packages/distributed/core.py#line=1017), in send_recv
    response = await comm.read(deserializers=deserializers)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
[…]
