### 2025-11-19 #8332 @platipodium
Just checking the status, these three are still open for me to continue the review

- [x] https://github.com/pangeo-fish/pangeo-fish/issues/227
- [ ] https://github.com/pangeo-fish/pangeo-fish/issues/226
- [x] https://github.com/pangeo-fish/pangeo-fish/issues/229
