### 2025-11-26 pangeo-fish/pangeo-fish#229 @platipodium
Thanks, works for me now.  You might want to give a hint to users and describe this phenomenon (in a line in the tutorial).
