### 2025-11-26 pangeo-fish/pangeo-fish#230 @platipodium
Improve zenodo metadata
Related https://github.com/openjournals/joss-reviews/issues/8332

It's great to see the development on GitHub mirrored and archived on Zenodo.  Zenodo's big problem for reproducible science is its lack of metadata enforcement.  I recommend that you create a `.zenodo.json` file with rich metadata that can be used to automatically fill many of the Zenodo fields upon upload there.
