### 2025-11-26 pangeo-fish/pangeo-fish#231 @platipodium
Add legal aspect to contributing guidelines
Related https://github.com/openjournals/joss-reviews/issues/8332, #226, #227

Thank you for clarifying how to contribute.  Your current statement covers appreciation and technical details, but it lacks legal information.  Answer the question: If I contribute to your software, under what terms will my contribution be part of that software?

You can find guidance on how this could be resolved in https://www.harmonyagreements.org (or others).
