### 2025-11-26 pangeo-fish/pangeo-fish#232 @platipodium
Fix documentation formatting
Related https://github.com/openjournals/joss-reviews/issues/8332

* There are some markdown issues on https://pangeo-fish.readthedocs.io/en/latest/notebook.html.
* Add a link to HEALPix format for users that are not familiar with it.
