### 2025-11-26 pangeo-fish/pangeo-fish#231 @annefou
Would that work to add:

## Licensing of Contributions
This project is licensed under the GNU General Public License v3.0 (see [LICENSE](https://github.com/pangeo-fish/pangeo-fish/issues/LICENSE)).
By submitting a contribution, you agree that:
- Your contribution is licensed under GPL v3.0
- You have the right to submit it under this license
