### 2025-11-26 pangeo-fish/pangeo-fish#231 @tinaok
Thanks @annefou , if we go for EUPL, would that be similar?
