### 2025-12-01 382623c
correct fish histogram
