### 2025-12-01 249be0e
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
