+++ pangeo_fish/bathy.py
@@ -407,10 +407,9 @@ def compute_pdf_bathy_batch_numpy(ds_chunk, reshaped_tag, copernicus_chunk):
     prof_corr = pressure[:, None, :] - XE[:, :, None] + 10  # (T, C, O)
     valid = ~np.isnan(prof_corr)
 
-    depth_bins = ds_chunk.depth_bins.values            
-    depth_bin_size = depth_bins[1] - depth_bins[0]     
-    
- 
+    depth_bins = ds_chunk.depth_bins.values
+    depth_bin_size = depth_bins[1] - depth_bins[0]
+
     bin_index = np.floor(prof_corr / depth_bin_size).astype("int32")
     bin_index[~valid] = 0
     max_bin = hist.shape[1] - 1
