### 2025-12-02 19d9960
Fix ruff warnings (unused imports, ambiguous variable, unused vars)
