### 2025-12-02 5f51456
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
