### 2025-12-02 4f490da
dealing with new conv method in helpers
