### 2025-12-02 91dcff7
adapt helpers
