+++ pangeo_fish/helpers.py
@@ -1200,8 +1200,8 @@ def optimize_pdf(
 
 def predict_positions(
     *,
-    ds: xr.Dataset | None = None,      
-    target_root: str | None = None,    
+    ds: xr.Dataset | None = None,
+    target_root: str | None = None,
     storage_options: dict,
     chunks: dict,
     track_modes=["mean", "mode"],
@@ -1243,24 +1243,23 @@ def predict_positions(
     """
 
     if ds is None:
-            if target_root is None:
-                raise ValueError(
-                    "You must provide either `ds` or `target_root` "
-                    "to load the dataset."
-                )
-    
-            # old behavior preserved:
-            emission = xr.open_dataset(
-                f"{target_root}/combined.zarr",
-                engine="zarr",
-                chunks=chunks,
-                inline_array=True,
-                storage_options=storage_options,
+        if target_root is None:
+            raise ValueError(
+                "You must provide either `ds` or `target_root` " "to load the dataset."
             )
-            emission = emission.compute()
-    
+
+        # old behavior preserved:
+        emission = xr.open_dataset(
+            f"{target_root}/combined.zarr",
+            engine="zarr",
+            chunks=chunks,
+            inline_array=True,
+            storage_options=storage_options,
+        )
+        emission = emission.compute()
+
     else:
-        
+
         emission = ds
 
         emission = emission.compute()
