### 2025-12-02 2e5c86f
proposition for using copernicusmarine
