### 2025-12-03 #8332 @tinaok
Thank you @platipodium and @RichardLitt. We’re actively working with our teams to resolve the remaining copyright and licensing questions, and I’ll update you as soon as we have everything clarified.
