### 2026-01-12 #8332 @RichardLitt
Thanks for your response, @platipodium. I appreciate it. You're right, I had not fully considered the reviewers' feelings here. I apologise. I am a new editor, and have much to learn.
