### 2026-01-19 a82642c
TIDE: update notebooks and add tide functions
