### 2026-01-19 e784564
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
