+++ notebooks/Tide/tide_function.py
@@ -1,4 +1,4 @@
-import xarray as xr 
+import xarray as xr
 import pandas as pd
 from ipywidgets import interact, IntSlider
 import numpy as np
@@ -14,11 +14,9 @@ from tqdm import tqdm
 from scipy.stats import multivariate_normal
 
 
-
 #####################################################################
-    #  Première partie: Extraction signal de marée par sinus
+#  Première partie: Extraction signal de marée par sinus
 #####################################################################
-
 
 
 def lssinfit(time, depth):
@@ -26,43 +24,50 @@ def lssinfit(time, depth):
     Ajuste un modèle sinusoïdal simple sur la profondeur.
     Retourne rmse, rsquare, amplitude.
     """
-    
-    depth_smoothed = savgol_filter(depth, window_length=15, polyorder=2, mode='mirror')
-    
+
+    depth_smoothed = savgol_filter(depth, window_length=15, polyorder=2, mode="mirror")
+
     w = 2 * np.pi / (12.42 / 24)  # fréquence de marée (rad/jour)
     X = np.column_stack([np.ones_like(time), np.sin(w * time), np.cos(w * time)])
-    #print(depth_smoothed, X)
-    #print(len(time),len(depth_smoothed))
+    # print(depth_smoothed, X)
+    # print(len(time),len(depth_smoothed))
     coef, _, _, _ = lstsq(X, depth_smoothed, rcond=None)
     fitted = X @ coef
     residuals = depth_smoothed - fitted
-    
+
     ss_res = np.sum(residuals**2)
-    ss_tot = np.sum((depth - np.mean(depth))**2)
+    ss_tot = np.sum((depth - np.mean(depth)) ** 2)
     rsq = 1 - ss_res / ss_tot if ss_tot > 0 else np.nan
     rmse = np.sqrt(np.mean(residuals**2))
-    amplitude = np.sqrt(coef[1]**2 + coef[2]**2)
-    return rmse, rsq, amplitude #, depth_smoothed
+    amplitude = np.sqrt(coef[1] ** 2 + coef[2] ** 2)
+    return rmse, rsq, amplitude  # , depth_smoothed
 
 
-def tide_behav_extr(df, tagno,
-                    tideFL=10, tideLV=[0.5, 0.7, 0.5],
-                    behavFL=16, behavLV=[0.5, 0.7, 0.5],
-                    dt=10):
+def tide_behav_extr(
+    df,
+    tagno,
+    tideFL=10,
+    tideLV=[0.5, 0.7, 0.5],
+    behavFL=16,
+    behavLV=[0.5, 0.7, 0.5],
+    dt=10,
+):
     """
     Traduction Python de la fonction MATLAB tidebehavextr.m
     Entrée : CSV avec colonnes 'time', 'depth' (et éventuellement 'temp')
     Sortie : xarray.Dataset avec résultats de classification marée / comportement
     """
     # --- Lecture des données brutes ---
-    
-    if 'time' not in df.columns or 'pressure' not in df.columns:
+
+    if "time" not in df.columns or "pressure" not in df.columns:
         raise ValueError("Le CSV doit contenir au moins 'time' et 'pressure'")
-    df['time'] = pd.to_datetime(df['time'])
-    df = df.sort_values('time')
-    time = (df['time'] - df['time'].iloc[0]).dt.total_seconds() / 3600.0 / 24.0  # en jours
-    depth = df['pressure'].to_numpy()
-    print(df['time'] - df['time'].iloc[0])
+    df["time"] = pd.to_datetime(df["time"])
+    df = df.sort_values("time")
+    time = (
+        (df["time"] - df["time"].iloc[0]).dt.total_seconds() / 3600.0 / 24.0
[…]
