### 2026-01-19 80d5598
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
