+++ notebooks/Wave/Wave_function.py
@@ -1,5 +1,5 @@
 ###########
-#Usual libraries
+# Usual libraries
 ###########
 import numpy as np
 import matplotlib.pyplot as plt
@@ -12,13 +12,15 @@ import pandas as pd
 from scipy.fft import rfft, rfftfreq
 from scipy.interpolate import interp1d
 from scipy.signal import fftconvolve
+
 ###########
-#Signal processing functions
+# Signal processing functions
 ###########
-from scipy.signal import detrend,welch 
+from scipy.signal import detrend, welch
 import xarray as xr
+
 ###########
-#Additionals functions
+# Additionals functions
 ###########
 from scipy.io import loadmat
 from scipy.stats.distributions import chi2
@@ -27,72 +29,81 @@ from IPython import display
 from scipy.signal import fftconvolve, welch
 import hvplot.xarray
 
+
 def stations_proches(ds, lat, lon, delta_lat=0.2, delta_lon=0.2):
     lat_min, lat_max = lat - delta_lat, lat + delta_lat
     lon_min, lon_max = lon - delta_lon, lon + delta_lon
 
     mask_nodes = (
-        (ds["latitude"] >= lat_min) & (ds["latitude"] <= lat_max) &
-        (ds["longitude"] >= lon_min) & (ds["longitude"] <= lon_max)
+        (ds["latitude"] >= lat_min)
+        & (ds["latitude"] <= lat_max)
+        & (ds["longitude"] >= lon_min)
+        & (ds["longitude"] <= lon_max)
     )
     index_nodes = ds.node.values[mask_nodes]  # Les vrais IDs
 
-    nodes = ds.sel(node=index_nodes)          # Sélection par valeur, pas par position
+    nodes = ds.sel(node=index_nodes)  # Sélection par valeur, pas par position
     return nodes, index_nodes
+
 
 def station_proche(ds, lat, lon):
     dlat = ds["latitude"] - lat
     dlon = ds["longitude"] - lon
-    
+
     # distance approximative en degrés (valable localement)
     dist2 = dlat**2 + dlon**2
 
-def conv_time(date_debut, date_evt, pas, step='seconds'):
-    
+
+def conv_time(date_debut, date_evt, pas, step="seconds"):
+
     # Calcul de la différence de temps
     diff_t = date_evt - date_debut
     # Conversion en secondes (ou autre unité selon step)
-    if step == 'seconds':
+    if step == "seconds":
         diff_unit = diff_t.total_seconds()
-    elif step == 'minutes':
+    elif step == "minutes":
         diff_unit = diff_t.total_seconds() / 60
-    elif step == 'hours':
+    elif step == "hours":
         diff_unit = diff_t.total_seconds() / 3600
-    elif step == 'days':
-        diff_unit = diff_t.total_seconds() / (3600*24)
+    elif step == "days":
+        diff_unit = diff_t.total_seconds() / (3600 * 24)
     else:
-        raise ValueError("Unité de temps non supportée. Choisir 'seconds', 'minutes', 'hours' ou 'days'")
-    
+        raise ValueError(
+            "Unité de temps non supportée. Choisir 'seconds', 'minutes', 'hours' ou 'days'"
+        )
+
     # Calcul de l'indice
     indice = int(diff_unit / pas)
-    
+
     return indice
-    
+
     idx = dist2.argmin()
     print(idx)
     return ds.isel(node=idx)
 
-def dispNewtonTH(f,dep):
-# This function inverts the linear dispersion relation (2*pi*f)^2=g*k*tanh(k*dep) to get 
[…]
