### 2026-03-09 7938f19
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/pre-commit-hooks: v5.0.0 → v6.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v5.0.0...v6.0.0)
- [github.com/psf/black-pre-commit-mirror: 25.1.0 → 26.3.0](https://github.com/psf/black-pre-commit-mirror/compare/25.1.0...26.3.0)
- [github.com/keewis/blackdoc: v0.4.0 → v0.4.6](https://github.com/keewis/blackdoc/compare/v0.4.0...v0.4.6)
- [github.com/astral-sh/ruff-pre-commit: v0.11.13 → v0.15.5](https://github.com/astral-sh/ruff-pre-commit/compare/v0.11.13...v0.15.5)
- [github.com/kynan/nbstripout: 0.8.1 → 0.9.1](
[…]
