### 2026-03-09 c60e67a
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
