### 2026-03-10 4786e5b
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
