### 2026-03-10 7e2643b
Fix license and copyright issue (#238)
* Create .mailmap for contributor identity mapping #219

Add contributor identities for canonicalization.

* Add files via upload

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix license, copyright, contribution, and Zenodo metadata

* fix test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre
[…]
