### 2026-03-10 #8332 @tinaok
@platipodium

Thank you for the helpful suggestions.

I have addressed the licensing and metadata issues in the repository:

- Unified the repository license metadata to GPL-3.0-only.
- Updated inconsistent license mentions in the README, pyproject.toml, and codemeta.json.
- Added a top-level COPYRIGHT file clarifying the copyright holders.
- Added a .zenodo.json file so that Zenodo releases will include complete metadata (authors, license, etc.).
- Added a section in CONTRIBUTING.md clarifying that contributions are licensed under the same license as the project.

These changes were implemented in the following pull request and have now been merged:

https://github.com/pangeo-fish/pangeo-fish/pull/238

Please let me know if anything further should be adjusted.

And thank you @RichardLitt for your patience!
