+++ docs/environment.yaml
@@ -56,6 +56,7 @@ dependencies:
   - dask-image
   - jupyterlab
   - ipywidgets
+  - sphinx_copybutton
   - pip
   - pip:
       - nbsphinx_link>=1.3.1
