+++ docs/environment.yaml
@@ -56,7 +56,7 @@ dependencies:
   - dask-image
   - jupyterlab
   - ipywidgets
-  - sphinx_copybutton
+  - sphinx-copybutton
   - pip
   - pip:
       - nbsphinx_link>=1.3.1
