+++ docs/environment.yaml
@@ -57,6 +57,7 @@ dependencies:
   - jupyterlab
   - ipywidgets
   - sphinx-copybutton
+  - myst-parser
   - pip
   - pip:
       - nbsphinx_link>=1.3.1
