### 2026-03-10 cf0c914
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
