+++ docs/environment.yaml
@@ -55,11 +55,11 @@ dependencies:
   - copernicusmarine
   - dask-image
   - jupyterlab
-  - docutils<0.21  # ADD THIS LINE
+  - docutils<0.21 # ADD THIS LINE
   - ipywidgets
-  - sphinx-remove-toctrees  # ADD THIS
-  - sphinx-book-theme  # ADD THIS
-  - nbsphinx           # ADD THIS
+  - sphinx-remove-toctrees # ADD THIS
+  - sphinx-book-theme # ADD THIS
+  - nbsphinx # ADD THIS
   #- sphinx-copybutton
   #- myst-parser
   - pip
