### 2026-04-08 pangeo-fish/pangeo-fish#232 @sschoombie
The link to the HEALPix documentation is currently a broken link. Would be good to fix.
