### 2026-04-13 pangeo-fish/pangeo-fish#232 @capetienne
Hi @sschoombie , I opened a PR to fix the broken HEALPix link: #251.
