### 2026-04-13 #8332 @capetienne
Hi @sschoombie,

Thanks for addressing the broken link issue. I confirm that it has been resolved in https://github.com/pangeo-fish/pangeo-fish/pull/251.

Best regards.
