### 2026-04-21 #8332 @RichardLitt
@tinaok @capetienne Can you make a new tagged release? See the "Additional author tasks" above, or https://joss.readthedocs.io/en/latest/editing.html#post-review.
