### 2026-04-28 13327b6
Grammar fixes (#253)
* Adding note saying it is OK to only cite the top level dep

* Grammar fixes

I tried to make this read a bit easier. Let me know what you think.
