+++ publication/paper.bib
@@ -321,6 +321,7 @@ oftware{siu_kwan_lam_2024_10839385,
   url          = {https://doi.org/10.5281/zenodo.10839385}
 }
 
+% Note: The editors have requested that we only cite top-level dependencies here, so we are not citing sparse and opt_einsum.
 % TODO: figure out if we need to cite sparse and opt_einsum: this is used by `xarray-healpy` (which, in turn, would need to cite `sparse` and `opt_einsum`). I'd opt for not doing so, but not sure.
 % sparse: no preferred citation
 % toolz: no preferred citation
