+++ publication/paper.md
@@ -48,61 +48,57 @@ bibliography: paper.bib
 # Summary
 
 Geo-referenced data plays an important role in understanding and conserving natural resources, particularly when investigating biological phenomena such as fish migration and habitat uses.
-Biologging, the practice of attaching small devices (called _tags_) to animals for recording behavior, physiology, and environmental data, proves to be invaluable in this field.
+Biologging, the practice of attaching small devices (called _tags_) to animals for recording behavior, physiology, and environmental data, is invaluable in this field.
 
 As fish can not be tracked directly using tracking devices such as GPS receivers, geolocation models have emerged to estimate fish positions by correlating individual series of physical measurements — e.g. temperature and pressure records — with geophysical reference fields — oceanic temperature and bathymetry — derived from satellite observations and hydrodynamical model outputs.
 
-Beside the difficulty of working with vast earth science datasets (due to their size and diversity), there is no open source implementation for biologged fish tracking.
-Yet, these fish geolocation models are critical for better understanding fish behavior and are nowadays seen as a powerful tool by policy makers to improve fish management and conservation.
+Besides the difficulty of working with vast earth science datasets (due to their size and diversity), there is no open source implementation for biologged fish tracking.
+Yet, these fish geolocation models are critical for understanding fish behavior better and are currently seen as a powerful tool by policy makers to improve fish management and conservation.
 
 To address this challenge, we developed a Python package, named **pangeo-fish**, for <u>fish tracking estimation</u>.
-It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source environment for interactive data analysis in the field of big data marine and geoscience.
+It is based on the [Pangeo](https://www.pangeo.io/) ecosystem, which offers a unique interoperable, scalable, open source environment for interactive data analysis in the fields of marine big data and geoscience.
 
 # Statement of need
 
-Biologging consists in attaching onto (or sometimes inserting into) an animal an
-electronic device that will record in its memory physical and/or geochemical parameters as a function of time so that scientists can reconstruct the activity of the animal, the characteristics of the environment it travels in and the interactions between the two.
+Biologging consists of attaching an electronic device onto (and sometimes inserting into) an animal, which will record in its memory physical and/or geochemical parameters as a function of time so that scientists can reconstruct the activity of the animal, the characteristics of the environment it travels in and the interactions between the two.
 
-These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, such as diving and activity patterns, energy use and interaction with environment.
+These tools can provide a wealth of information on the behaviors and movements of free-swimming marine animals, such as diving and activity patterns, energy use and environmental interactions.
 
 ![Promotion of the FISH-INTEL tagging campaign.\label{fig:fishintel}](fishintel.png){ width=35% }
 
-However, unlike terrestrial animals or marine mammals, whose positions can be directly estimated using ARGOS or GPS technologies, tracking fish underwater is challenging.
-To address this issue, various tagging experiments have been conducted on a variety of fish species [@spanish_tagging; @skagerrak_tagging], and methods have been proposed for approximating the fish locations, referred to as geolocation models [@pontual_seabass_migration_2023; @woillez_hmm-based_2016].
+However, unlike terrestrial and some marine mammals, whose positions can be directly estimated using ARGOS or GPS technologies, tracking fish underwater is challenging.
+To address this issue, various tagging experiments have been conducted on a variety of fish species [@spanish_tagging; @skagerrak_tagging], and methods have been proposed for approximating fish locations, referred to as geolocation models [@pontual_seabass_migration_2023; @woillez_hmm-based_2016].
 
[…]
