### 2026-04-30 #8332 @capetienne
Hi @RichardLitt,
Archived release is made today here :

https://github.com/pangeo-fish/pangeo-fish/releases/tag/2026.04.0

Thank you!
