### 2026-05-01 a377f97
fix missing attribute
