### 2026-05-03 94e1180
fix: make pip install -e . self-sufficient + numpy/xarray compat fixes (#256)
* fix: make pip install -e . self-sufficient + numpy/xarray compat fixes

- pyproject.toml: declare all runtime deps the package code imports
  (matplotlib, hvplot, holoviews, cartopy, cmocean, geopandas, shapely,
  toolz, distributed, flox, tqdm, rich, rich-click, dask, pint,
  intake-xarray, h5netcdf, kerchunk, gcsfs, numcodecs)
- pyproject.toml: pin known-working IO stack
  (zarr==2.18.7, kerchunk==0.2.7, fsspec==2025.3.2, s3fs==2025.3.2,
  gcsfs==2025.3.2, numcodecs==0.15.1)
- pyproject.toml: cap xarray<2025.4 (r
[…]
