### 2026-05-04 dfbbe80
feat(light): add solar geolocation sub-package
Adds pangeo_fish.light with solar threshold-based geolocation:
- ingest.py: load_tag_csv / prepare_tag_folder for lotek, wc_psat, dst formats
- quality.py: dynamic_threshold, twilight_quality, compute_quality_flags
- solar.py: detect_twilight_events, self_calibrate_solar_threshold, compute_solar_likelihood
- unit tests for dynamic_threshold and compute_solar_likelihood output dims
- ephem added to CI environment; pyproject.toml restored from upstream main
