+++ pangeo_fish/light/solar.py
@@ -1,0 +1,556 @@
+"""Solar threshold geolocation method (Part A of the light pipeline).
+
+Detects twilight events (RISE / SET) by threshold-crossing, self-calibrates
+the solar elevation threshold from the first nights of deployment, evaluates
+event quality, and computes per-night solar likelihood maps on a regular
+lat/lon grid.
+"""
+
+from typing import NamedTuple
+
+import ephem
+import numpy as np
+import pandas as pd
+import xarray as xr
+from scipy.ndimage import gaussian_filter
+
+from pangeo_fish.light.quality import _tz
+
+# ---------------------------------------------------------------------------
+# Private helpers
+# ---------------------------------------------------------------------------
+
+
+def _get_astro_prediction(date, lon, lat, horizon_deg=-10.0):
+    """Return predicted astronomical RISE and SET for *date* at *(lon, lat)*.
+
+    Parameters
+    ----------
+    date : pd.Timestamp or datetime.date
+        Calendar date.
+    lon : float
+        Observer longitude (degrees East).
+    lat : float
+        Observer latitude (degrees North).
+    horizon_deg : float, default -10.0
+        Solar elevation defining "rise" and "set" (degrees).
+
+    Returns
+    -------
+    rise : pd.Timestamp or None
+    sett : pd.Timestamp or None
+        UTC timestamps of the predicted rise and set.  Both are ``None``
+        if :mod:`ephem` cannot find a crossing (e.g., polar day/night).
+    """
+    obs = ephem.Observer()
+    obs.lon = str(lon)
+    obs.lat = str(lat)
+    obs.pressure = 0
+    obs.horizon = str(horizon_deg)
+    sun = ephem.Sun()
+    try:
+        obs.date = f"{date.year}/{date.month:02d}/{date.day:02d} 00:00:00"
+        rise = pd.Timestamp(
+            ephem.Date(obs.next_rising(sun, use_center=True)).datetime()
+        ).tz_localize("UTC")
+        obs.date = f"{date.year}/{date.month:02d}/{date.day:02d} 10:00:00"
+        sett = pd.Timestamp(
+            ephem.Date(obs.next_setting(sun, use_center=True)).datetime()
+        ).tz_localize("UTC")
+        return rise, sett
+    except Exception:
+        return None, None
+
+
+def _get_crossing_fast(t_approx, lon, lat, horizon_deg, direction="up", window_h=6.0):
+    """Return the predicted solar crossing near *t_approx* for a grid pixel.
+
+    Parameters
+    ----------
+    t_approx : pd.Timestamp
+        Approximate observed crossing time (UTC).
+    lon : float
+        Observer longitude (degrees East).
+    lat : float
+        Observer latitude (degrees North).
+    horizon_deg : float
+        Solar elevation defining the crossing (degrees).
+    direction : {'up', 'down'}, default 'up'
+        ``'up'`` for RISE, ``'down'`` for SET.
+    window_h : float, default 6.0
+        Search half-window (hours).
+
+    Returns
+    -------
+    pd.Timestamp or None
+        Predicted crossing time (UTC), or ``None`` if not found within
+        ``window_h`` of ``t_approx``.
+
+    Notes
+    -----
+    **Why ``window_h=6.0`` (was 3.0):** at high-latitude grid pixels in
+    summer, the predicted crossing can shift by several hours relative to
+    the observed time at mid-latitudes.  The old 3-hour window caused
+    silent ``None`` returns → single-event fallback → ghost likelihood
+    peaks.  6 hours is wide enough to capture any real crossing while
+    remaining tighter than a full 12-hour day/night cycle.
+    """
+    obs = ephem.Observer()
+    obs.lon = str(lon)
+    obs.lat = str(lat)
[…]
