### 2026-05-04 1b8e652
feat(diff,bathy,io,pdf): likelihood and pipeline improvements
- diff.py: extract Ch2 likelihood functions from notebook; add likelihood_fast_gu (numba guvectorize)
- bathy.py: remove dynamic_bathymetry; fix XE squeeze for singleton depth dim; pass storage_options in batch_compute_pdf_bathy
- helpers.py: add decode_healpix() xdggs/xarray compat fallback
- io.py: open_tag accepts tagging_event.csv and tagging_events.csv variants; handle tz-naive timestamps
- pdf.py: combine_emission_pdf_with_light — remove pdf_moon param; dynamic zarr filename from emission components
- tests: update test_diff a
[…]
