### 2026-05-04 6b31244
feat(notebooks): new_workflow/ pipeline — multi-tag ingestion, variance, solar light, README
- new_workflow/: production notebooks reorganised in ordered pipeline
  - raw_to_dst.ipynb: one-shot conversion + S3 upload (lotek/wc_psat/dst)
  - prepare_tag_model.ipynb: model download via copernicusmarine
  - variance_fast.ipynb: depth-dependent temperature variance from Argo
  - pangeo_fish_with_light.ipynb: main pipeline (Ch1–Ch8), lunar chapter removed
- new_workflow/README.md: file format spec (dst.csv, tagging_events.csv, metadata.json), tag types table, pipeline order
- notebooks/README.md: u
[…]
