### 2026-05-04 4d46d58
feat(light): add lunar template-fit PDF chapter
Adds pangeo_fish.light lunar pipeline on top of the solar sub-package:
- physics.py: moon_ground_illuminance + predicted_moon_curve
- lunar.py: detect_moon_events, compute_moon_likelihood (Spearman correlation fit)
- __init__.py: export lunar and physics symbols
- pdf.py: restore pdf_moon=None parameter in combine_emission_pdf_with_light
- pangeo_fish_with_light.ipynb: restore Ch7 lunar template-fit PDF cells
- tests: restore test_moon_illuminance_* and test_correlation_to_likelihood_normalization
- pyproject.toml: add ephem dependency for lunar 
[…]
