+++ pangeo_fish/light/lunar.py
@@ -1,0 +1,342 @@
+"""Lunar template-fit geolocation method (Part B of the light pipeline).
+
+Identifies usable moon nights, computes Spearman correlation between
+the observed light curve and the physically predicted moonlight curve at
+each grid pixel, converts correlations to likelihoods, and stacks the
+results into an xarray DataArray ready for HEALPix regridding.
+"""
+
+import time as tmod
+
+import ephem
+import numpy as np
+import pandas as pd
+import xarray as xr
+from scipy.stats import spearmanr
+
+from pangeo_fish.light.physics import predicted_moon_curve
+
+
+def find_usable_moon_nights(
+    df,
+    min_phase=60,
+    max_depth_m=15,
+    bin_min=15,
+    min_bins=8,
+    min_samples_per_bin=3,
+):
+    """Identify calendar nights that are usable for lunar geolocation.
+
+    A night is usable if:
+
+    * Mean moon phase > ``min_phase`` (bright enough).
+    * At least 30 shallow observations (``pressure < max_depth_m``).
+    * At least ``min_bins`` time bins of ``bin_min`` minutes, each with
+      ≥ ``min_samples_per_bin`` samples.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        Tag data with columns ``time`` (UTC), ``light``, ``pressure``.
+    min_phase : float, default 60
+        Minimum moon illuminated fraction (%) at midnight.
+    max_depth_m : float, default 15
+        Maximum depth (m / dbar) considered "shallow".
+    bin_min : int, default 15
+        Duration of each time bin (minutes).
+    min_bins : int, default 8
+        Minimum number of valid bins required per night.
+    min_samples_per_bin : int, default 3
+        Minimum number of raw samples per bin.
+
+    Returns
+    -------
+    list of dict
+        Each dict has keys:
+        ``date`` (pd.Timestamp), ``phase`` (float, %),
+        ``times`` (np.ndarray of bin-centre datetimes),
+        ``light`` (np.ndarray of binned median light counts),
+        ``n_bins`` (int), ``n_raw`` (int).
+    """
+    all_dates = pd.date_range(
+        df["time"].min().normalize(),
+        df["time"].max().normalize(),
+        freq="D",
+        tz="UTC",
+    )
+    moon_nights = []
+
+    for d in all_dates:
+        # Check moon phase at ~01:00 UTC (representative of the night)
+        obs_chk = ephem.Observer()
+        obs_chk.pressure = 0
+        obs_chk.date = (d + pd.Timedelta(hours=1)).strftime("%Y/%m/%d %H:%M:%S")
+        phase = float(ephem.Moon(obs_chk).phase)
+
+        if phase < min_phase:
+            continue
+
+        # Night window: 19:00 UTC on day d to 08:00 UTC on day d+1 (32 h span)
+        t0 = d + pd.Timedelta(hours=19)
+        t1 = d + pd.Timedelta(hours=32)
+
+        mask = (df["time"] >= t0) & (df["time"] <= t1) & (df["pressure"] < max_depth_m)
+        sub = df[mask].copy()
+        if len(sub) < 30:
+            continue
+
+        # Bin into ``bin_min``-minute medians
+        sub["bin"] = sub["time"].dt.floor(f"{bin_min}min")
+        binned = (
+            sub.groupby("bin")
+            .agg(light_med=("light", "median"), n=("light", "count"))
+            .reset_index()
+        )
+        binned = binned[binned["n"] >= min_samples_per_bin]
+
+        if len(binned) < min_bins:
+            continue
+
+        moon_nights.append(
+            {
+                "date": d,
+                "phase": phase,
[…]
