### 2026-05-05 47e951c
feat(light): add wc_archival and wc_psat_light ingestion + LightLoc pipeline
ingest.py:
- add tag_type="wc_archival": reads out-Archive.csv (15 s, Light Level column)
- add tag_type="wc_psat_light": reads *-Series.csv + *-LightLoc.csv (Transmission)
- new ingest_lightloc(): standardises WC LightLoc CSV → lightloc.csv in tag folder
- prepare_tag_folder: new lightloc_csv_path param, writes lightloc.csv for wc_psat_light

solar.py:
- new pairs_from_lightloc(): builds (pairs, qdf) from lightloc.csv, same format
  as detect_twilight_events + compute_quality_flags — depth filter via max_depth

__ini
[…]
