+++ pangeo_fish/light/ingest.py
@@ -9,6 +9,7 @@ by :func:`pangeo_fish.io.open_tag`:
         dst.csv              ← time, temperature, pressure, light
         tagging_event.csv   ← copied from the source file
         metadata.json        ← tag_name, tag_type
+        lightloc.csv         ← (optional) WC crossing times for PSAT-light tags
 
 Supported tag types
 -------------------
@@ -25,6 +26,20 @@ Supported tag types
     (``*-Series.csv``) with columns ``Day``, ``Time``, ``Depth``,
     ``Temperature``.  No raw light counts are available, so ``light``
     is set to NaN and ``HAS_LIGHT`` must be ``False``.
+
+``"wc_archival"``
+    Wildlife Computers MiniPAT recovered archival tag — high-resolution
+    Archive CSV (``out-Archive.csv``) with 15-second records and a
+    ``Light Level`` column.  Produces ``dst.csv`` with ``light`` filled.
+    Use ``HAS_LIGHT = True`` and ``LIGHT_SOURCE = "series"``.
+
+``"wc_psat_light"``
+    Wildlife Computers MiniPAT non-recovered PSAT that transmitted
+    twilight crossing times via Argos.  Requires both the 10-minute
+    Series CSV (``*-Series.csv``) for temperature/depth and the LightLoc
+    CSV (``*-LightLoc.csv``) for crossing times.  Produces ``dst.csv``
+    (``light = NaN``) **and** ``lightloc.csv``.
+    Use ``HAS_LIGHT = True`` and ``LIGHT_SOURCE = "lightloc"``.
 """
 
 import json
@@ -42,7 +57,7 @@ def load_tag_csv(path, tag_type):
     ----------
     path : str or path-like
         Path to the CSV file.
-    tag_type : {"lotek", "wc_psat"}
+    tag_type : {"lotek", "wc_psat", "wc_archival", "dst"}
         Manufacturer / format identifier.
 
     Returns
@@ -87,7 +102,7 @@ def load_tag_csv(path, tag_type):
             }
         )[["temperature", "pressure", "light"]]
 
-    elif tag_type == "wc_psat":
+    elif tag_type in ("wc_psat", "wc_psat_light"):
         df_raw = pd.read_csv(path)
         df_raw["time"] = pd.to_datetime(
             df_raw["Day"] + " " + df_raw["Time"], format="%d-%b-%Y %H:%M:%S"
@@ -105,6 +120,25 @@ def load_tag_csv(path, tag_type):
             }
         )
 
+    elif tag_type == "wc_archival":
+        # out-Archive.csv: "HH:MM:SS DD-Mon-YYYY", columns Time/Depth/Temperature/Light Level
+        df_raw = pd.read_csv(path)
+        df_raw["time"] = pd.to_datetime(
+            df_raw["Time"].str.strip(), format="%H:%M:%S %d-%b-%Y", errors="coerce"
+        )
+        df_raw = (
+            df_raw.dropna(subset=["time", "Depth", "Temperature"])
+            .sort_values("time")
+            .set_index("time")
+        )
+        dst = pd.DataFrame(
+            {
+                "temperature": df_raw["Temperature"],
+                "pressure": df_raw["Depth"],
+                "light": df_raw["Light Level"],
+            }
+        )
+
     elif tag_type == "dst":
         # Standard pangeo-fish dst.csv format — already processed
         dst = pd.read_csv(path, parse_dates=["time"], index_col="time").sort_index()
@@ -112,11 +146,84 @@ def load_tag_csv(path, tag_type):
 
     else:
         raise ValueError(
-            f"Unknown tag_type {tag_type!r}. Supported: 'lotek', 'wc_psat', 'dst'."
+            f"Unknown tag_type {tag_type!r}. "
[…]
