### 2026-05-05 117d496
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
