### 2026-05-16 6286944
Fix grammatical errors in paper.md
