### 2026-05-16 11a8c0c
Format titles in paper.bib for consistency
