+++ publication/paper.bib
@@ -404,7 +404,7 @@ doi = {10.1080/10236244.2021.1912604}
 }
 
 @article{pedersen2008geolocation,
-  title={Geolocation of North Sea cod (Gadus morhua) using hidden Markov models and behavioural switching},
+  title={Geolocation of {North} {Sea} cod (Gadus morhua) using hidden Markov models and behavioural switching},
   author={Pedersen, Martin W{\ae}ver and Righton, David and Thygesen, Uffe H{\o}gsbro and Andersen, Ken Haste and Madsen, Henrik},
   journal={Canadian Journal of Fisheries and Aquatic Sciences},
   volume={65},
@@ -426,7 +426,7 @@ doi = {10.1080/10236244.2021.1912604}
 }
 
 @article{righton2010thermal,
-  title={Thermal niche of Atlantic cod Gadus morhua: limits, tolerance and optima},
+  title={Thermal niche of {Atlantic} cod {Gadus} morhua: limits, tolerance and optima},
   author={Righton, David A and Andersen, Ken Haste and Neat, Francis and Thorsteinsson, Vilhjalmur and Steingrund, Petur and Sved{\"a}ng, Henrik and Michalsen, Kathrine and Hinrichsen, Hans-Harald and Bendall, Victoria and Neuenfeldt, Stefan and others},
   journal={Marine Ecology Progress Series},
   volume={420},
