### 2026-05-16 bedd090
Fix formatting in titles of bibliography entries
