### 2026-05-16 a217a25
Update affiliation for Anne Fouilloux (#260)
* Update affiliation for Anne Fouilloux

* Update acknowledgements in README.md
