### 2026-05-20 d1cc7cf
add minimum version constraint for copernicusmarine (>=2.4.1)
