### 2026-05-22 493679b
fix: update dependency versions in pyproject.toml
- Relax xarray upper bound to >=2026.4.0
- Upgrade zarr to >=3.1.6 (v3)
- Upgrade kerchunk to >=0.2.9
- Add copernicusmarine==2.3.0
- Pin distributed==2025.2.0
