### 2026-05-22 64f3331
fix: update dependency pins and fix cf_xarray 0.10.10 bounds compatibility
- Relax fsspec/s3fs/gcsfs pins to >=2025.9.0 to include HTTP parquet fix
- Relax numcodecs pin to >=0.15.1
- Bump dask and distributed to >=2026.3.0 to fix 8x slowdown with zarr3+kerchunk
- Fix count_detections in acoustic.py: cf_xarray >=0.10.10 returns both axis
  alias ('T') and coordinate name ('time') as separate keys in cf.bounds,
  causing len(cf.bounds)==2 and a false ValueError. Now counts unique bound
  variable names instead of keys.

Co-Authored-By: capetienne <etienne.cap98@gmail.com>
