### 2026-05-22 9fe7e75
fix: remove unused import of first after acoustic.py refactor
Co-Authored-By: capetienne <etienne.cap98@gmail.com>
