### 2026-05-22 3f39802
feat(helpers): default load_model to Copernicus Marine when uri is None
- _open_copernicus_model now takes a copernicus_model_name and fetches
  thetao/zos + matching _static (deptho/mask) via copernicusmarine.
- load_model: uri becomes optional; when None, use copernicus_model_name
  (default "cmems_mod_glo_phy_my_0.083deg_P1D-m").
- Old intake YAML loader renamed to _open_intake_catalog.
- pin copernicusmarine>=2.3,<2.4 in pyproject and 2.3.* in conda envs.

Refs #234.
