+++ pangeo_fish/helpers.py
@@ -6,6 +6,7 @@ import warnings
 from importlib.metadata import version
 from pathlib import Path
 
+import copernicusmarine
 import fsspec
 import holoviews as hv
 import imageio as iio
@@ -334,12 +335,12 @@ def plot_tag(
     return plot
 
 
-def _open_copernicus_model(yaml_url: str, chunks: dict = None):
+def _open_intake_catalog(yaml_url: str, chunks: dict = None):
     """Open an intake catalog.
 
     Parameters
     ----------
-    catalog_url : str
+    yaml_url : str
         Path to the ``.yaml`` file
     chunks : dict, optional
         How to chunk the data
@@ -352,6 +353,68 @@ def _open_copernicus_model(yaml_url: str, chunks: dict = None):
     cat = intake.open_catalog(yaml_url)
     model = open_copernicus_catalog(cat, chunks)
     return model
+
+
+def _open_copernicus_model(
+    copernicus_model_name: str,
+    bbox: dict[str, tuple[float, float]],
+    time_slice: slice,
+    tag_log: xr.Dataset = None,
+):
+    """Open a Copernicus Marine dataset and merge it with its static fields.
+
+    Parameters
+    ----------
+    copernicus_model_name : str
+        Copernicus Marine ``dataset_id`` for the daily-mean physics product
+        (e.g. ``"cmems_mod_glo_phy_my_0.083deg_P1D-m"``). The matching static
+        product (``..._static``) is opened to retrieve ``deptho`` and ``mask``.
+    bbox : dict
+        Spatial boundaries with keys ``longitude`` and ``latitude``.
+    time_slice : slice
+        Time slice used to bound ``start_datetime`` / ``end_datetime``. If
+        ``None``, the full extent of ``tag_log`` is used.
+    tag_log : xarray.Dataset, optional
+        DST data; used as fallback when ``time_slice`` is ``None``.
+
+    Returns
+    -------
+    model : xarray.Dataset
+        Dataset with ``TEMP``, ``XE`` and ``H0`` (via :func:`prepare_dataset`).
+    """
+    static_name = copernicus_model_name.rsplit("_", 1)[0] + "_static"
+
+    if time_slice is not None and time_slice.start is not None:
+        start_datetime = pd.Timestamp(time_slice.start).strftime("%Y-%m-%d")
+        end_datetime = pd.Timestamp(time_slice.stop).strftime("%Y-%m-%d")
+    elif tag_log is not None:
+        start_datetime = np.datetime_as_string(tag_log.time.values[0], unit="D")
+        end_datetime = np.datetime_as_string(tag_log.time.values[-1], unit="D")
+    else:
+        raise ValueError("Either `time_slice` or `tag_log` must be provided.")
+
+    ds_thetao_zos = copernicusmarine.open_dataset(
+        dataset_id=copernicus_model_name,
+        variables=["thetao", "zos"],
+        minimum_longitude=bbox["longitude"][0],
+        maximum_longitude=bbox["longitude"][1],
+        minimum_latitude=bbox["latitude"][0],
+        maximum_latitude=bbox["latitude"][1],
+        start_datetime=start_datetime,
+        end_datetime=end_datetime,
+    )
+    static_var = copernicusmarine.open_dataset(
+        dataset_id=static_name,
+        variables=["deptho", "mask"],
+        minimum_longitude=bbox["longitude"][0],
+        maximum_longitude=bbox["longitude"][1],
+        minimum_latitude=bbox["latitude"][0],
+        maximum_latitude=bbox["latitude"][1],
+    )
+    static_var = static_var.assign_coords(longitude=ds_thetao_zos["longitude"].values)
+    ds_all = xr.merge([ds_thetao_zos, static_var], compat="no_conflicts")
[…]
