### 2026-05-22 77d6797
fix(helpers): accept ndarray time_slice + update tutorial notebook to uri=None
