### 2026-05-22 59f5ca9
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
