### 2026-06-03 a8b83d3
Fix formatting of abstract and title in paper.bib
