+++ publication/paper.bib
@@ -7,7 +7,7 @@
     pages = {501-515},
     year = {2018},
     month = {07},
-    abstract = {From 2010 to 2012, 246 data storage tags were deployed on European seabass in the Iroise Natural Marine Park, a marine protected area (MPA) off west Brittany, France. A return rate of 14.6\% associated with long time series of data provided new information on fish ecology (e.g. maximum experienced depth greater than 225 m, temperature range 6.80–21.87°C). Depth and temperature series were used to infer individual migration using an innovative hidden Markov model (HMM) especially developed for seabass geolocation. Reconstructed fish tracks revealed that seabass is a partially migratory species, as individuals exhibited either long-distance migrations towards the Bay of Biscay or the Celtic Sea, or residency behaviour in the Iroise Sea. Fidelity to summer feeding areas and to winter spawing areas was demonstrated. These results suggest that the population is spatially structured. The Iroise Sea is likely a mixing zone for different stocks or sub-populations, and may also shelter a resident population. At the population scale, such findings may impact ICES stock assessment and the resulting decisions from EU managers. At the local scale, conservation action could be taken by MPA managers. Besides, this study demonstrates the high potential of archival tags for investigating multi-year behavioural patterns such as site fidelity to offshore spawning areas.},
+    abstract = {From 2010 to 2012, 246 data storage tags were deployed on European seabass in the Iroise Natural Marine Park, a marine protected area (MPA) off west Brittany, France. A return rate of 14.6\% associated with long time series of data provided new information on fish ecology (e.g. maximum experienced depth greater than 225 m, temperature range 6.80–21.87°C). Depth and temperature series were used to infer individual migration using an innovative hidden {Markov} model (HMM) especially developed for seabass geolocation. Reconstructed fish tracks revealed that seabass is a partially migratory species, as individuals exhibited either long-distance migrations towards the Bay of Biscay or the Celtic Sea, or residency behaviour in the Iroise Sea. Fidelity to summer feeding areas and to winter spawing areas was demonstrated. These results suggest that the population is spatially structured. The Iroise Sea is likely a mixing zone for different stocks or sub-populations, and may also shelter a resident population. At the population scale, such findings may impact ICES stock assessment and the resulting decisions from EU managers. At the local scale, conservation action could be taken by MPA managers. Besides, this study demonstrates the high potential of archival tags for investigating multi-year behavioural patterns such as site fidelity to offshore spawning areas.},
     issn = {1054-3139},
     doi = {10.1093/icesjms/fsy086},
     url = {https://doi.org/10.1093/icesjms/fsy086},
@@ -404,7 +404,7 @@ doi = {10.1080/10236244.2021.1912604}
 }
 
 @article{pedersen2008geolocation,
-  title={Geolocation of {North} {Sea} cod ({Gadus} morhua) using hidden Markov models and behavioural switching},
+  title={Geolocation of {North} {Sea} cod ({Gadus} morhua) using hidden {Markov} models and behavioural switching},
   author={Pedersen, Martin W{\ae}ver and Righton, David and Thygesen, Uffe H{\o}gsbro and Andersen, Ken Haste and Madsen, Henrik},
   journal={Canadian Journal of Fisheries and Aquatic Sciences},
   volume={65},
