### 2026-06-03 07aa338
Format keywords and title with proper casing
