+++ publication/paper.bib
@@ -37,7 +37,7 @@
     volume = {22},
     number = {5},
     pages = {1041-1066},
-    keywords = {archival tag, Atlantic halibut, geolocation accuracy and precision, Gulf of St. Lawrence, Hidden Markov Model, pop-up satellite archival tag},
+    keywords = {archival tag, Atlantic halibut, geolocation accuracy and precision, Gulf of St. Lawrence, Hidden {Markov} Model, pop-up satellite archival tag},
     doi = {10.1111/faf.12568},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12568},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/faf.12568},
@@ -436,7 +436,7 @@ doi = {10.1080/10236244.2021.1912604}
 }
 
 @article{heerah2017coupling,
-  title={Coupling spectral analysis and hidden Markov models for the segmentation of behavioural patterns},
+  title={Coupling spectral analysis and hidden {Markov} models for the segmentation of behavioural patterns},
   author={Heerah, Karine and Woillez, Mathieu and Fablet, Ronan and Garren, Fran{\c{c}}ois and Martin, St{\'e}phane and De Pontual, H{\'e}l{\`e}ne},
   journal={Movement Ecology},
   volume={5},
